### 2023-05-12 #5463 @editorialbot
[PRE REVIEW]: The Spreadsheet Energy System Model Generator (SESMG): A tool for the optimization of urban energy systems
**Submitting author:** @chrklemm (<a href="http://orcid.org/0000-0003-0801-4178">Christian Klemm</a>)
**Repository:** https://github.com/SESMG/SESMG
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.0
