### 2023-05-16 f5b905f
docu(installation.rst): add note for python version input
