### 2023-05-16 b6c6353
docu(installation.rst): fix notes
