### 2023-05-17 9520d8f
docu(trouble_shooting.rst): add python version installation issue
