+++ requirements.txt
@@ -19,4 +19,5 @@ seaborn>=0.11.0
 dash>=2.4.0  
 dash_canvas>=0.1.0 
 h5py==3.6.0
+pyutilib>=6.0.0
 pytest
