### 2023-05-18 682a483
docu(structure_of_energy_systems.rst): remove caps from load profile names
