### 2023-05-18 7c5d8ee
docu(model_definition.rst): typo
