### 2023-05-19 e7f9ae6
docu(trouble_shooting.rst): add windpower issue to M-12, fix code components M23-M27
