### 2023-05-19 1994462
docu(installation.rst): note about incompatibility with py 3.9.7
