+++ docs/02.01.00_installation.rst
@@ -11,6 +11,10 @@ Windows users can finde the versions here: `Python download page for windows <ht
 .. note:: 
 
 	Make sure to use the installer referring to your windows version (e.g. "Windows x86-64 executable installer" for windows 64-bit version). You can check the build in bit version following `these steps <https://support.microsoft.com/en-us/windows/32-bit-and-64-bit-windows-frequently-asked-questions-c6ca9541-8dce-4d48-0415-94a3faa2e13d>`_.
+	
+.. note:: 
+
+	The SESMG is not compatible with the python version 3.9.7, due to a confliciting bug in this version to the subpackage streamlit. You need do install another pyhton version or update to another python version e. g. 3.9.13, which is widely tested.
 
 .. figure:: images/manual/Installation/sesmg_installation_ms_1.png
    :width: 50 %
@@ -117,7 +121,10 @@ Go to the `Python download page for macOS <https://www.python.org/downloads/maco
 .. note:: 
 
 	Make sure to use the installer referring to your chip (Intel-only or universal2) version (e.g. "macOS 64-bit universal2 installer" for mac with Apple M1 / M2). You can check the kind of buil in chip by opening the "About this Mac" option behind the apple icon in the above left corner.
+	
+.. note:: 
 
+	The SESMG is not compatible with the python version 3.9.7, due to a confliciting bug in this version to the subpackage streamlit. You need do install another pyhton version or update to another python version e. g. 3.9.13, which is widely tested.
 
 .. figure:: images/manual/Installation/sesmg_installation_ms_1.png
    :width: 50 %
@@ -199,6 +206,10 @@ Linux
 1.1 Make sure a Python 3 Version is installed:
 
 ``$ python3 --version``
+
+.. note:: 
+
+	The SESMG is not compatible with the python version 3.9.7, due to a confliciting bug in this version to the subpackage streamlit. You need do install another pyhton version or update to another python version e. g. 3.9.13, which is widely tested.
 	 
 2. Download the Spreadsheet Energy System Model Generator from `GitHub <https://github.com/SESMG/SESMG>`_ as .zip folder and extract the .zip folder into any directory on the computer.
 
