### 2023-05-19 446a845
docu(trouble_shooting.rst): add py 3.9.7 error
