+++ program_files/preprocessing/pre_model_analysis.py
@@ -5,8 +5,9 @@ import pandas
 import logging
 
 
-def filter_result_component_types(components: pandas.DataFrame,
-                                  component_type: str) -> pandas.DataFrame:
+def filter_result_component_types(
+    components: pandas.DataFrame, component_type: str
+) -> pandas.DataFrame:
     """
         returns dataframe containing only one specific component type
         
@@ -25,9 +26,12 @@ def filter_result_component_types(components: pandas.DataFrame,
 
 
 def update_component_investment_decisions(
-        components: pandas.DataFrame, model_definition_path: str,
-        model_definition_type_name: str, result_type_name: str,
-        investment_boundary_factor: int, investment_boundaries=True
+    components: pandas.DataFrame,
+    model_definition_path: str,
+    model_definition_type_name: str,
+    result_type_name: str,
+    investment_boundary_factor: int,
+    investment_boundaries=True,
 ) -> (pandas.DataFrame, list):
     """
         Adapts investment decision depending on the results of a \
@@ -64,52 +68,54 @@ def update_component_investment_decisions(
     """
     # return components type results
     result_components = filter_result_component_types(
-        components=components,
-        component_type=result_type_name)
+        components=components, component_type=result_type_name
+    )
 
     # read the component type model definition sheet
     components_xlsx = pandas.read_excel(
-        io=model_definition_path,
-        sheet_name=model_definition_type_name)
+        io=model_definition_path, sheet_name=model_definition_type_name
+    )
     # drop first (nan) column
     components_xlsx = components_xlsx.iloc[1:, :]
     # reset of index required, so that it is uniform to the result-dataframe
     components_xlsx = components_xlsx.reset_index()
-    
+
     component_type_switch_dict = {
         "district heating": dh_technical_pre_selection,
         "buses": bus_technical_pre_selection,
-        "insulation": insulation_technical_pre_selection
+        "insulation": insulation_technical_pre_selection,
     }
-    
+
     try:
         # run the component specific method if applicable in the
         # component_type_switch_dict
         component_type_switch_dict.get(model_definition_type_name)(
-            components_xlsx=components_xlsx,
-            result_components=result_components)
+            components_xlsx=components_xlsx, result_components=result_components
+        )
 
         # return the updated components_xlsx file as well as an empty
         # list since no components were deactivated
         return components_xlsx, []
-    
+
     except TypeError:
         # run the general technical pre selection if the component_type
         # is not applicable within the component_type_switch_dict
         list_of_deactivated_components = technical_pre_selection(
-            components_xlsx=components_xlsx,
-            result_components=result_components)
+            components_xlsx=components_xlsx, result_components=result_components
+        )
         if investment_boundaries:
             tightening_investment_boundaries(
                 components_xlsx=components_xlsx,
                 result_components=result_components,
-                investment_boundary_factor=investment_boundary_factor)
-    
+                investment_boundary_factor=investment_boundary_factor,
+            )
+
         return components_xlsx, list_of_deactivated_components
 
 
-def technical_pre_selection(components_xlsx: pandas.DataFrame,
-                            result_components: pandas.DataFrame) -> list:
+def technical_pre_selection(
+    components_xlsx: pandas.DataFrame, result_components: pandas.DataFrame
+) -> list:
     """
         deactivates investment-components for which no investments has \
         been carried out and additionally returns a list of deactivated\
[…]
