+++ tests/test_preprocessing_create_graph.py
@@ -1,10 +1,8 @@
 def test_linebreaks():
     from program_files.preprocessing.create_graph import linebreaks
-
     pass
 
 
 def test_ESGraphRenderer():
     from program_files.preprocessing.create_graph import ESGraphRenderer
-
     pass
