+++ program_files/preprocessing/components/Transformer.py
@@ -77,14 +77,13 @@ class Transformers:
         :param nodes: list of components created before (can be empty)
         :type nodes: list
     """
-
+    
     # intern variables
     nodes_transformer = []
     busd = None
-
-    def get_primary_output_data(
-        self, transformer: pandas.Series, max_invest=None
-    ) -> dict:
+    
+    def get_primary_output_data(self, transformer: pandas.Series,
+                                max_invest=None) -> dict:
         """
             Within this method the output flow data of the currently
             considered transformer is parameterized and returned to the
@@ -125,33 +124,31 @@ class Transformers:
         # create the transformers investment object which is set as
         # investment flow afterwards
         invest = Investment(
-            ep_costs=transformer["periodical costs"],
-            minimum=transformer["min. investment capacity"],
-            maximum=transformer["max. investment capacity"],
-            periodical_constraint_costs=transformer["periodical constraint costs"],
-            existing=transformer["existing capacity"],
-            nonconvex=True if transformer["non-convex investment"] == 1 else False,
-            offset=transformer["fix investment costs"],
-            fix_constraint_costs=transformer["fix investment constraint costs"],
-        )
+                ep_costs=transformer["periodical costs"],
+                minimum=transformer["min. investment capacity"],
+                maximum=transformer["max. investment capacity"],
+                periodical_constraint_costs=transformer[
+                    "periodical constraint costs"],
+                existing=transformer["existing capacity"],
+                nonconvex=True if transformer["non-convex investment"] == 1
+                else False,
+                offset=transformer["fix investment costs"],
+                fix_constraint_costs=transformer[
+                    "fix investment constraint costs"])
         # return the parameterized flow to the main creation process
         return {
             "outputs": {
                 self.busd[transformer["output"]]: Flow(
-                    variable_costs=transformer["variable output costs"],
-                    emission_factor=transformer["variable output constraint costs"],
-                    investment=invest,
+                        variable_costs=transformer["variable output costs"],
+                        emission_factor=transformer[
+                            "variable output constraint costs"],
+                        investment=invest,
                 )
             }
         }
-
-    def create_transformer(
-        self,
-        transformer: pandas.Series,
-        inputs: dict,
-        conversion_factors: dict,
-        outputs: dict,
-    ) -> None:
+    
+    def create_transformer(self, transformer: pandas.Series, inputs: dict,
+                           conversion_factors: dict, outputs: dict) -> None:
         """
             Within this method the parameterized transformer (inputs,
             outputs and conversion factors already created) is added to
@@ -176,13 +173,15 @@ class Transformers:
             :type conversion_factors: dict
         """
         self.nodes_transformer.append(
-            Transformer(
-                label=transformer["label"], **inputs, **outputs, **conversion_factors
-            )
+                Transformer(label=transformer["label"],
+                            **inputs,
+                            **outputs,
+                            **conversion_factors)
         )
         logging.info("\t Transformer created: " + transformer["label"])
-
-    def generic_transformer(self, transformer: pandas.Series, two_input=False) -> None:
+    
+    def generic_transformer(self, transformer: pandas.Series,
+                            two_input=False) -> None:
         """
             Creates a generic transformer with the parameters given in
             'nodes_data' and adds it to the list of components 'nodes'.
@@ -229,81 +228,68 @@ class Transformers:
         }
         # Defines Capacity values for the second transformer output
         if transformer["output2"] not in ["None", "none", 0]:
-            efficiency_ratio = float(transformer["efficiency2"]) / float(
-                transformer["efficiency"]
-            )
+            efficiency_ratio = float(transformer["efficiency2"]) \
[…]
