+++ tests/test_preprocessing_averaging.py
@@ -1,4 +1,4 @@
+
 def test_timeseries_averaging():
-    from program_files.preprocessing.data_preparation_algorithms.averaging import (
-        timeseries_averaging,
-    )
+    from program_files.preprocessing.data_preparation_algorithms.averaging \
+        import timeseries_averaging
