### 2023-05-19 f15a136
tests(create_energy_system): add a new test
