### 2023-05-22 4e307bc
installation(python version): remove python 3.7 as possible python version since oemof solph is not longer python 3.7 compatible
