### 2023-05-22 2cd726d
tests(bus, link): update tests to oemof solph 0.5.0 folder structure
