### 2023-05-22 2e24d27
tests(bus, link): update to oemof.solph 0.5.0 structure
