### 2023-05-22 354dfbe
ci(JOSS): reduce workflow on push on the master branch pushes
