+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,8 @@
-on: [push]
-
+on:
+  push:
+    branches:
+      - master
+      
 jobs:
   paper:
     runs-on: ubuntu-latest
