### 2023-05-22 e43460d
code(create_energy_system): bug fix
