+++ program_files/preprocessing/create_energy_system.py
@@ -124,7 +124,7 @@ def define_energy_system(nodes_data: dict) -> (EnergySystem, dict):
                                        freq=temp_resolution)
     
     # initialisation of the energy system
-    esys = EnergySystem(timeindex=datetime_index, infer_last_interval=True)
+    esys = EnergySystem(timeindex=datetime_index, infer_last_interval=False)
     # setting the index column for time series and weather data
     for sheet in ["timeseries", "weather data"]:
         # defines a time series
