### 2023-05-23 b3ec136
code(US-Tool): move the length of geothermal probe to the us-sheet, since this is a location-dependent variable.
