+++ docs/02.02.03_urban_district_upscaling.rst
@@ -106,27 +106,27 @@ Category 3
 
 
 .. csv-table:: Input for the upscaling sheet. Category 3: Central investment data.
-   :header:  label,comment,active,technology,latitude,longitude,area,dh_connection,azimuth,surface tilt,flow temperature
+   :header:  label,comment,active,technology,latitude,longitude,area,dh_connection,azimuth,surface tilt,flow temperature, length of the geoth. probe
 
-   ,,,,° WGS 84,° WGS 84,sqm,,°,°,°C
-   electricity_exchange,,1,electricity_exchange,,,,,,, 
-   battery_storage,,1,battery,,,,,,, 
-   ng_chp,,0,naturalgas_chp,,,,heat_input,,, 
-   bg_chp,,0,biogas_chp,,,,heat_input,,, 
-   pe_chp,,0,pellet_chp,,,,heat_input,,, 
-   wc_chp,,1,woodchips_chp,,,,heat_input,,, 
-   swhp,,0,swhp_transformer,,,,heat_input,,, 
-   ashp,,0,ashp_transformer,,,,heat_input,,, 
-   gchp,free area needed,1,gchp_transformer,,,2500,heat_input,,, 
-   ng_heating,,0,naturalgas_heating_plant,,,,heat_input,,, 
-   bg_heating,,0,biogas_heating_plant,,,,heat_input,,, 
-   pe_heating,,0,pellet_heating_plant,,,,heat_input,,, 
-   wc_heating,,1,woodchips_heating_plant,,,,heat_input,,, 
-   thermal_storage,,1,thermal_storage,,,,heat_input,,, 
-   p2g,,0,power_to_gas,,,,heat_input,,, 
-   heat_input,heat center,1,heat_input_bus,52,7,,,,,40
-   central_pv_st,free area needed,1,pv&st,52,7,15000,heat_input,180,22.5, 
-   screw_turbine,,1,timeseries_source,,,,,,, 
+   ,,,,° WGS 84,° WGS 84,sqm,,°,°,°C,m
+   electricity_exchange,,1,electricity_exchange,,,,,,,, 
+   battery_storage,,1,battery,,,,,,,, 
+   ng_chp,,0,naturalgas_chp,,,,heat_input,,,, 
+   bg_chp,,0,biogas_chp,,,,heat_input,,,, 
+   pe_chp,,0,pellet_chp,,,,heat_input,,,, 
+   wc_chp,,1,woodchips_chp,,,,heat_input,,,, 
+   swhp,,0,swhp_transformer,,,,heat_input,,,, 
+   ashp,,0,ashp_transformer,,,,heat_input,,,, 
+   gchp,free area needed,1,gchp_transformer,,,2500,heat_input,,,,100
+   ng_heating,,0,naturalgas_heating_plant,,,,heat_input,,,, 
+   bg_heating,,0,biogas_heating_plant,,,,heat_input,,,, 
+   pe_heating,,0,pellet_heating_plant,,,,heat_input,,,, 
+   wc_heating,,1,woodchips_heating_plant,,,,heat_input,,,, 
+   thermal_storage,,1,thermal_storage,,,,heat_input,,,, 
+   p2g,,0,power_to_gas,,,,heat_input,,,, 
+   heat_input,heat center,1,heat_input_bus,52,7,,,,,40,
+   central_pv_st,free area needed,1,pv&st,52,7,15000,heat_input,180,22.5,, 
+   screw_turbine,,1,timeseries_source,,,,,,,, 
 
 
 - **label**: The technology name can be chosen arbitrarily by the user and represents the ID of a central technology. The ID must be unique for each technology, because all following columns are assigned to it.
@@ -139,6 +139,7 @@ Category 3
 - **azimuth	(°)**: For ground-mounted solar systems, the azimuth must be specified. 
[…]
