+++ tests/test_urban_district_upscaling_central_components.py
@@ -70,6 +70,7 @@ def test_create_power_to_gas_entry():
                             "None",
                             "central_test_heat_bus"],
                 "area": [float(0)] * 3,
+                "length of the geoth. probe": [0.0] * 3,
                 "temperature high": ["0"]* 3}),
             right=transformers,
             on="transformer_type").drop(columns=["transformer_type"]),
@@ -191,6 +192,7 @@ def test_central_heating_plant_entry():
                  "output": ["test_output_bus"],
                  "output2": ["None"],
                  "area": [float(0)],
+                 "length of the geoth. probe": [0.0],
                  "temperature high": "0"}),
             right=transformers,
             on="transformer_type").drop(columns=["transformer_type"]),
@@ -249,6 +251,7 @@ def test_central_CHP_entry():
                      "output": ["central_test_elec_bus"],
                      "output2": ["test_output_bus"],
                      "area": [float(0)],
+                     "length of the geoth. probe": [0.0],
                      "temperature high": "0"}),
                 right=transformers,
                 on="transformer_type").drop(columns=["transformer_type"]),
