### 2023-05-23 dd37796
code(US-Tool): move the heat extraction to the us-sheet, since this is a location-dependent variable.
