+++ program_files/urban_district_upscaling/components/Central_components.py
@@ -9,7 +9,7 @@ import pandas
 def create_central_heat_component(
         label: str, comp_type: str, bus: str, exchange_buses: dict,
         sheets: dict, area: str, standard_parameters: pandas.ExcelFile,
-        flow_temp: str, len_geoth_probe: str) -> dict:
+        flow_temp: str, len_geoth_probe: str, heat_extraction: str) -> dict:
     """
         In this method, all heat supply systems are calculated for a
         heat input into the district heat network.
@@ -38,6 +38,9 @@ def create_central_heat_component(
         :param len_geoth_probe: length of the vertical heat exchanger \
             relevant for GCHPs
         :type len_geoth_probe: str
+        :param heat_extraction: heat extraction for the heat exchanger \
+            referring to the location
+        :type heat_extraction: str
         
         :return: - **sheets** (dict) - dictionary containing the \
             pandas.Dataframes that will represent the model \
@@ -97,7 +100,8 @@ def create_central_heat_component(
             area=area,
             standard_parameters=standard_parameters,
             flow_temp=flow_temp,
-            len_geoth_probe=len_geoth_probe
+            len_geoth_probe=len_geoth_probe,
+            heat_extraction=heat_extraction
         )
         # increase indicator to prevent duplex bus creation
         central_heatpump_indicator += 1
@@ -277,6 +281,9 @@ def central_comp(central: pandas.DataFrame, true_bools: list, sheets: dict,
                         len_geoth_probe=comp["length of the geoth. probe"]
                         if comp["technology"] == "gchp_transformer"
                         else "0",
+                        heat_extraction=comp["heat extraction"]
+                        if comp["technology"] == "gchp_transformer"
+                        else "0",
                         standard_parameters=standard_parameters,
                         flow_temp=bus["flow temperature"]
                     )
@@ -410,7 +417,8 @@ def create_central_heatpump(label: str, specification: str, create_bus: bool,
                             central_electricity_bus: bool, output: str,
                             sheets: dict, area: str,
                             standard_parameters: pandas.ExcelFile,
-                            flow_temp: str, len_geoth_probe: str) -> dict:
+                            flow_temp: str, len_geoth_probe: str,
+                            heat_extraction: str) -> dict:
     """
          In this method, a central heatpump unit is created, for this
          purpose the necessary data set is obtained
@@ -445,6 +453,9 @@ def create_central_heatpump(label: str, specification: str, create_bus: bool,
         :param len_geoth_probe: length of the vertical heat exchanger \
             relevant for GCHPs
         :type len_geoth_probe: str
+        :param heat_extraction: heat extraction for the heat exchanger \
+            referring to the location
+        :type heat_extraction: str
         
         :return: - **sheets** (dict) - dictionary containing the \
             pandas.Dataframes that will represent the model \
@@ -482,7 +493,8 @@ def create_central_heatpump(label: str, specification: str, create_bus: bool,
         area=area,
         standard_parameters=standard_parameters,
         flow_temp=flow_temp,
-        len_geoth_probe=len_geoth_probe
+        len_geoth_probe=len_geoth_probe,
+        heat_extraction=heat_extraction
     )
 
 
