+++ tests/test_urban_district_upscaling_central_components.py
@@ -71,6 +71,7 @@ def test_create_power_to_gas_entry():
                             "central_test_heat_bus"],
                 "area": [float(0)] * 3,
                 "length of the geoth. probe": [0.0] * 3,
+                "heat extraction": [0.0] * 3,
                 "temperature high": ["0"]* 3}),
             right=transformers,
             on="transformer_type").drop(columns=["transformer_type"]),
@@ -139,6 +140,7 @@ def test_central_heatpump_entry():
                 "output2": ["None"],
                 "area": [100.0],
                 "length of the geoth. probe": [100.0],
+                "heat extraction": [0.0328],
                 "temperature high": ["60"]}),
             right=transformers,
             on="transformer_type").drop(columns=["transformer_type"]),
@@ -161,7 +163,8 @@ def test_create_central_heatpump(test_central_heatpump_entry):
         area="100",
         standard_parameters=standard_parameters,
         len_geoth_probe="100",
-        flow_temp="60"
+        flow_temp="60",
+        heat_extraction="0.0328"
     )
     
     for key in sheets.keys():
@@ -193,6 +196,7 @@ def test_central_heating_plant_entry():
                  "output2": ["None"],
                  "area": [float(0)],
                  "length of the geoth. probe": [0.0],
+                 "heat extraction": [0.0],
                  "temperature high": "0"}),
             right=transformers,
             on="transformer_type").drop(columns=["transformer_type"]),
@@ -252,6 +256,7 @@ def test_central_CHP_entry():
                      "output2": ["test_output_bus"],
                      "area": [float(0)],
                      "length of the geoth. probe": [0.0],
+                     "heat extraction": [0.0],
                      "temperature high": "0"}),
                 right=transformers,
                 on="transformer_type").drop(columns=["transformer_type"]),
