### 2023-05-23 3f82a86
code(US-Tool): this commit enables the entering of building specific electricity purchase prices
