+++ program_files/urban_district_upscaling/components/Bus.py
@@ -8,7 +8,7 @@ import pandas
 
 def create_standard_parameter_bus(label: str, bus_type: str, sheets: dict,
                                   standard_parameters: pandas.ExcelFile,
-                                  coords=None) -> dict:
+                                  coords=None, shortage_cost=None) -> dict:
     """
         Creates a bus with standard_parameters, based on the standard
         parameters given in the "standard_parameters" dataset and adds
@@ -28,6 +28,11 @@ def create_standard_parameter_bus(label: str, bus_type: str, sheets: dict,
         :param coords: latitude / longitude / dh column of the given bus\
             used to connect a producer bus to district heating network
         :type coords: list
+        :param shortage_cost: If the user wants to map a shortage \
+            price that differs from the standard parameter, this value
+            != None. This will overwrite the value from the standard
+            parameters.
+        :type shortage_cost: float
         
         :return: - **sheets** (dict) - dictionary containing the \
             pandas.Dataframes that will represent the model \
@@ -55,6 +60,10 @@ def create_standard_parameter_bus(label: str, bus_type: str, sheets: dict,
              "lat": coords[0], "lon": coords[1]})
     else:
         bus_dict.update({"district heating conn.": float(0)})
+    # If the user wants to map a shortage price that differs from the
+    # standard parameter
+    if shortage_cost is not None:
+        bus_dict.update({"shortage costs": shortage_cost})
     # appends the new created component to buses sheet
     return append_component(sheets, "buses", bus_dict)
 
@@ -122,6 +131,9 @@ def create_building_electricity_bus_link(
     """
     from program_files.urban_district_upscaling.components import Link
     pv_bus = check_rather_building_needs_pv_bus(building=building)
+    # get the users shortage costs input
+    shortage_cost = building["electricity cost"] \
+        if building["electricity cost"] != "standard" else None
     # create the building electricity bus if the building has a pv
     # system or gets an electricity demand later on
     if pv_bus or building["building type"] not in ["0", 0]:
@@ -130,7 +142,8 @@ def create_building_electricity_bus_link(
             label=(str(building["label"]) + "_electricity_bus"),
             bus_type=bus,
             sheets=sheets,
-            standard_parameters=standard_parameters
+            standard_parameters=standard_parameters,
+            shortage_cost=shortage_cost
         )
         # create link from central electricity bus to building
         # electricity bus if the central electricity exchange is enabled
