+++ tests/test_urban_district_upscaling_pre_processing.py
@@ -129,7 +129,8 @@ def test_create_heat_pump_buses_links(test_hp_buses_entry):
     building = {"label": "test",
                 "parcel ID": "test_parcel",
                 "ashp": "no",
-                "gchp": "yes"}
+                "gchp": "yes",
+                "heatpump electricity cost": "standard"}
     
     sheets = {"buses": pandas.DataFrame(),
               "links": pandas.DataFrame()}
@@ -216,7 +217,8 @@ def test_create_building_buses(test_building_buses_entry):
         "longitude": "0",
         "st 1": "yes",
         "pv 1": "yes",
-        "roof area 1": 10
+        "roof area 1": 10,
+        "electricity cost": "standard"
     }
     building.update({})
     
