### 2023-06-05 9980d24
code(US-Tool): add building specific electricity purchase emissions
