+++ docs/02.02.03_urban_district_upscaling.rst
@@ -44,12 +44,12 @@ Category 1
 ^^^^^^^^^^
 
 .. csv-table:: Input for the upscaling sheet. Category 1: Building-specific data.
-   :header:  label,comment,active,year of construction,distance of electric vehicles,electricity demand,heat demand,building type,units,occupants per unit,gross building area,latitude,longitude,year of construction wall,area outer wall,year of construction windows,area windows,year of construction roof,rooftype,area roof,cluster ID,flow temperature,electricity cost,heatpump electricity cost
+   :header:  label,comment,active,year of construction,distance of electric vehicles,electricity demand,heat demand,building type,units,occupants per unit,gross building area,latitude,longitude,year of construction wall,area outer wall,year of construction windows,area windows,year of construction roof,rooftype,area roof,cluster ID,flow temperature,electricity cost,heatpump electricity cost,electricity emission,heatpump electricity emission
 
-   x,,,,km/a,kWh / (sqm * a),kWh / (sqm * a),,,,sqm,° WGS 84,° WGS 84,,sqm,,sqm,,,sqm,,°C
-   001_building,,1,1800,0,400,400,COM_Food,1,1,100,52.000000,7.000000,1800,50,0,0,1967,flat roof,25,0,60,standard,standard
-   002_building,,1,1800,0,0,0,MFB,1,1,50,52.000000,7.000000,1979,100,1999,20,1993,flat roof,50,0,60,standard,standard
-   003_building,,1,1800,10000,30,20,SFB,1,1,120,52.000000,7.000000,1994,250,2001,125,1992,step roof,125,0,40,standard,standard
+   x,,,,km/a,kWh / (sqm * a),kWh / (sqm * a),,,,sqm,° WGS 84,° WGS 84,,sqm,,sqm,,,sqm,,°C,€/kWh,g/kWh,€/kWh,g/kWh
+   001_building,,1,1800,0,400,400,COM_Food,1,1,100,52.000000,7.000000,1800,50,0,0,1967,flat roof,25,0,60,standard,standard,standard,standard
+   002_building,,1,1800,0,0,0,MFB,1,1,50,52.000000,7.000000,1979,100,1999,20,1993,flat roof,50,0,60,standard,standard,standard,standard
+   003_building,,1,1800,10000,30,20,SFB,1,1,120,52.000000,7.000000,1994,250,2001,125,1992,step roof,125,0,40,standard,standard,standard,standard
    
 - **label**: The building name can be chosen by the user and is the identification number (ID) of a building. The ID must be unique for each building, because all the following columns are assigned to it.
 - **comment**: Space for an individual comment, e.g. an indication of which measure this component belongs to.
@@ -74,7 +74,9 @@ Category 1
 - **cluster ID**: The cluster ID is used to spatially assign a building to a specific area. The area can be, for example, a settlement or neighborhood. The cluster ID is crucial for spatial clustering.
 - **flow temperature	(°C)**: The flow temperature may differ depending on the heating system. The flow temperature should not fall below the heat source temperature of a heat pump. If the outdoor temperature is 35 °C and the flow temperature is 30 °C, the air heat pump is switched off and an alternative technology is used for heat supply.
 - **electricity cost (€/kWh)**: If the user wants to use a electricity purchase price that differs from the standard parameter (e.g. due to a green electricity tariff), this can be entered here. If not the user has to enter "standard".
-- **heatpump electricity cost (€/kWh): If the user wants to use a heatpump electricity purchase price that differs from the standard parameter (e.g. due to a different heatpump tariff), this can be entered here. If not the user has to enter "standard".
[…]
