### 2023-06-05 8ad5286
code(US-Tool): Implementation of the division of the sink for wood stove and St in order to represent their heat share correctly in the status quo.
