### 2023-06-05 42f9fda
GUI(urban_district_upscaling): redesign GUI page and varibales, add open fred input
