+++ program_files/GUI_st/pages/2_Urban_District_Upscaling.py
@@ -17,6 +17,9 @@ from program_files.GUI_st.GUI_st_global_functions \
 # settings the initial streamlit page settings
 st_settings_global()
 
+# opening the input value dict, which will be saved as a json
+GUI_udu_dict = {}
+
 # Import GUI help comments from the comment json and safe as an dict
 GUI_helper = import_GUI_input_values_json(
     os.path.dirname(os.path.dirname(__file__))
@@ -27,68 +30,116 @@ def us_application() -> None:
     """
         Definition of the sidebar elements for the urban district
         upscaling page and starting a udu tool run.
+
+        :returns: - **input_us_sheet** \
+                    (streamlit.runtime.uploaded_file_manager.UploadedFile) - \
+                    Streamlit varibale including the uploaded upscaling_sheet \
+                    (was a xlsx before)
+                - **input_standard_parameter** \
+                    (streamlit.runtime.uploaded_file_manager.UploadedFile) - \
+                    Streamlit varibale including the uploaded \
+                    standard_parameter_sheet (was a xlsx before)
     """
-    model_definition_df = ""
 
     # create form-submit element for multiple inputs
     with st.sidebar.form("Input Parameters"):
 
+        if "state_submitted_us_run" not in st.session_state:
+            st.session_state["state_submitted_us_run"] = False
+
         # input us sheet
-        input_us_sheet_path = st.file_uploader(
+        input_us_sheet = st.file_uploader(
             label="Import your upscaling sheet:",
             help=GUI_helper["udu_fu_us_sheet"])
 
         # input standard parameter sheet
-        input_standard_parameter_path = st.file_uploader(
+        input_standard_parameter = st.file_uploader(
             label="Import your standard parameter sheet:",
             help=GUI_helper["udu_fu_sp_sheet"])
 
         # text input to define the file name
-        result_file_name = \
+        GUI_udu_dict["result_file_name"] = \
             st.text_input(label="Type in your model definition file name.",
                           help=GUI_helper["udu_ti_model_def_name"])
 
+#TODO: add GUI help comments
+        # expander for weather data settings
+        with st.expander(label="Open Fred Weather Data"):
+
+            # input if open fres weather data should be added
+            GUI_udu_dict["input_open_fred"] = st.checkbox(
+                label="Download Open Fred Weather Data",
+                help=GUI_helper["udu_cb_weather_data"])
+
+            # coordinate input for weather data download
+            # longitude coordinates
+            GUI_udu_dict["input_cords_lon"] = st.text_input(
+                label="Longitude Coordinates",
+                help=GUI_helper["udu_ti_coords_lon"])
+            # latitude coordinates
+            GUI_udu_dict["input_cords_lat"] = st.text_input(
+                label="Latitude Coordinates",
+                help=GUI_helper["udu_ti_coords_lat"])
+
         # Submit button to start optimization.
-        submitted_us_run = st.form_submit_button(
+        st.session_state["state_submitted_us_run"] = st.form_submit_button(
                 label="Start US Tool",
                 help=GUI_helper["udu_fs_start_US_tool"])
 
-        # Run program main function if start button is clicked
-        if submitted_us_run:
-            if input_us_sheet_path != "" \
-                    and input_standard_parameter_path != "" \
-                    and result_file_name != "":
-                # strings replace due to variables defined above
-                us_path_list = [
-                    input_us_sheet_path,
-                    input_standard_parameter_path,
-                    result_file_name,
-                    os.path.join(
-                        os.path.dirname(__file__),
-                        r"../../urban_district_upscaling/plain_scenario.xlsx")
-                ]
-
-                model_definition_df = \
-                    urban_district_upscaling_pre_processing(
-                        paths=us_path_list,
[…]
