### 2023-06-05 4a519c1
tests(US-Tool): update tests to ensure compatibility to the new input variables
