+++ tests/test_urban_district_upscaling_sink.py
@@ -239,7 +239,9 @@ def test_create_heat_sink(test_SFB_heat_sink_entry,
             "building type": "SFB",
             "year of construction": 2000,
             "heat demand": 0,
-            "units": 1}),
+            "units": 1,
+            "wood stove share": "standard",
+            "solar thermal share": "standard"}),
         area=300,
         sheets={"sinks": pandas.DataFrame()},
         sinks_standard_param=sinks,
@@ -255,7 +257,8 @@ def test_create_heat_sink(test_SFB_heat_sink_entry,
             "building type": "COM_Food",
             "year of construction": 1980,
             "heat demand": 0,
-            "units": 1}),
+            "units": 1,
+            "wood stove share": "standard"}),
         area=300,
         sheets={"sinks": pandas.DataFrame()},
         sinks_standard_param=sinks,
@@ -271,7 +274,8 @@ def test_create_heat_sink(test_SFB_heat_sink_entry,
             "building type": "COM_Food",
             "year of construction": 0,
             "heat demand": 10,
-            "units": 0}),
+            "units": 0,
+            "wood stove share": "standard"}),
         area=300,
         sheets={"sinks": pandas.DataFrame()},
         sinks_standard_param=sinks,
@@ -346,7 +350,9 @@ def test_create_sinks(test_SFB_electricity_sink_entry,
             "occupants per unit": 3,
             "units": 1,
             "year of construction": 2000,
-            "distance of electric vehicles": 10000}),
+            "distance of electric vehicles": 10000,
+            "wood stove share": "standard",
+            "solar thermal share": "standard"}),
         sheets={"sinks": pandas.DataFrame(),
                 "time series": time_series},
         standard_parameters=standard_parameters)
