### 2023-06-05 906aa80
tests(US-Tool): update tests to ensure compatibility to the new input variables
