+++ tests/test_urban_district_upscaling_sink.py
@@ -258,7 +258,8 @@ def test_create_heat_sink(test_SFB_heat_sink_entry,
             "year of construction": 1980,
             "heat demand": 0,
             "units": 1,
-            "wood stove share": "standard"}),
+            "wood stove share": "standard",
+            "solar thermal share": "standard"}),
         area=300,
         sheets={"sinks": pandas.DataFrame()},
         sinks_standard_param=sinks,
@@ -275,7 +276,8 @@ def test_create_heat_sink(test_SFB_heat_sink_entry,
             "year of construction": 0,
             "heat demand": 10,
             "units": 0,
-            "wood stove share": "standard"}),
+            "wood stove share": "standard",
+            "solar thermal share": "standard"}),
         area=300,
         sheets={"sinks": pandas.DataFrame()},
         sinks_standard_param=sinks,
