### 2023-06-06 f790b87
documentation(US-Tool): add help comments for open fred in the GUI
