### 2023-06-06 0040150
tests(US-Tool): adapt test for new column in us sheet
