### 2023-06-07 27b0e8e
docu(trouble_shooting.rst): add osmx/cartopy warning
