### 2023-06-14 57f810a
Update 03.00.00_trouble_shooting.rst
