+++ program_files/urban_district_upscaling/clustering.py
@@ -1,4 +1,3 @@
-import pandas
 from program_files.urban_district_upscaling.components import (
     Link,
     Sink,
@@ -8,131 +7,60 @@ from program_files.urban_district_upscaling.components import (
     Source,
 )
 
-true_bools = ["yes", "Yes", 1, "1"]
 
+def clustering_transformers(
+    sheets, sink_parameters, transformer_parameters, cluster, standard_parameters
+):
+    """ """
 
-def clustering_transformers(sheets: dict, sink_parameters: list,
-                            transformer_parameters: dict, cluster: str,
-                            standard_parameters: pandas.ExcelFile) -> dict:
-    """
-        In this method the buses and links for the connection of the
-        clustered transformers as well as the clustered transformers
-        themselves are created and attached to the return data
-        structure "sheets".
-        
-        :param sheets: dictionary containing the pandas.Dataframes that\
-            will represent the model definition's Spreadsheets
-        :type sheets: dict
-        :param sink_parameters: list containing the cluster's sinks \
-            parameter (4) res_heat_demand, (5)com_heat_demand, \
-            (6) ind_heat_demand
-        :type sink_parameters: list
-        :param transformer_parameters: dict containing the cluster's \
-            transformer parameters (index) technology each entry is a \
-            list where index 0 is a counter
-        :type transformer_parameters: list
-        :param cluster: str containing the cluster's ID
-        :type cluster: str
-        :param standard_parameters: pandas imported ExcelFile \
-            containing the non-building specific technology data
-        :type standard_parameters: pandas.ExcelFile
-        
-        :return: - **sheets** (dict) - dictionary containing the \
-            pandas.Dataframes that will represent the model \
-            definition's Spreadsheets which was modified in this method
-    """
-    # check rather the cluster's heat demand is > 0
     if (sink_parameters[4] + sink_parameters[5] + sink_parameters[6]) != 0:
-        for technology in ["gasheating", "electricheating", "ashp", "gchp"]:
+        for i in ["gasheating", "electricheating", "ashp", "gchp"]:
             # create res or com gasheating
-            if transformer_parameters[technology][0] > 0:
+            if transformer_parameters[i][0] > 0:
                 # create cluster gas bus with building type weighted
                 # shortage costs
-                if technology == "gasheating":
+                if i == "gasheating":
                     sheets = Bus.create_cluster_averaged_bus(
-                        sink_parameters=sink_parameters,
-                        cluster=cluster,
-                        fuel_type="gas",
-                        sheets=sheets,
-                        standard_parameters=standard_parameters
+                        sink_parameters, cluster, "gas", sheets, standard_parameters
                     )
-                # create cluster hp electricity bus with building type \
-                # weighted shortage costs
-                elif technology in ["ashp", "gchp"] \
-                        and str(cluster) + "_gchp_building_link" not \
-                        in list(sheets["links"]["label"]):
+                elif (i == "ashp" or i == "gchp") and str(
+                    cluster
+                ) + "_gchp_building_link" not in list(sheets["links"]["label"]):
                     # create hp building type averaged price
                     sheets = Bus.create_cluster_averaged_bus(
-                        sink_parameters=sink_parameters,
-                        cluster=cluster,
-                        fuel_type="hp_elec",
-                        sheets=sheets,
-                        standard_parameters=standard_parameters
+                        sink_parameters, cluster, "hp_elec", sheets, standard_parameters
                     )
                     # electricity link from building electricity bus to hp
-                    # electricity bus
+                    # elec bus
                     sheets = Link.create_link(
                         label=str(cluster) + "_gchp_building_link",
                         bus_1=str(cluster) + "_electricity_bus",
[…]
