+++ program_files/urban_district_upscaling/urban_district_upscaling_post_processing.py
@@ -1,0 +1,288 @@
+# TODO decide whether include MAX PV and MAX ST or the percentage of area usage
+
+import pandas as pd
+import os
+
+
+def create_decentral_overview(components_csv_data):
+    # defining columns of the sheet including decentralized components
+    decentral_columns = [
+        "Building",
+        "PV 1",
+        "Max. PV 1",
+        "PV 2",
+        "Max. PV 2",
+        "PV 3",
+        "Max. PV 3",
+        "PV 4",
+        "Max. PV 4",
+        "PV 5",
+        "Max. PV 5",
+        "Installed PV",
+        "Max. PV",
+        "STC 1",
+        "Max. STC 1",
+        "STC 2",
+        "Max. STC 2",
+        "STC 3",
+        "Max. STC 3",
+        "STC 4",
+        "Max. STC 4",
+        "STC 5",
+        "Max. STC 5",
+        "Installed STC",
+        "Max. STC",
+        "Gasheating-System",
+        "ASHP",
+        "GCHP",
+        "Battery-Storage",
+        "Heat Transformer",
+        "Electric Heating",
+    ]
+    # defining units for decentral components
+    decentral_columns_units = {
+        "Building": "",
+        "PV 1": "(kW)",
+        "Max. PV 1": "(kW)",
+        "PV 2": "(kW)",
+        "Max. PV 2": "(kW)",
+        "PV 3": "(kW)",
+        "Max. PV 3": "(kW)",
+        "PV 4": "(kW)",
+        "Max. PV 4": "(kW)",
+        "PV 5": "(kW)",
+        "Max. PV 5": "(kW)",
+        "Installed PV": "(kW)",
+        "Max. PV": "(kW)",
+        "STC 1": "(kW)",
+        "Max. STC 1": "(kW)",
+        "STC 2": "(kW)",
+        "Max. STC 2": "(kW)",
+        "STC 3": "(kW)",
+        "Max. STC 3": "(kW)",
+        "STC 4": "(kW)",
+        "Max. STC 4": "(kW)",
+        "STC 5": "(kW)",
+        "Max. STC 5": "(kW)",
+        "Installed STC": "(kW)",
+        "Max. STC": "(kW)",
+        "Gasheating-System": "(kW)",
+        "ASHP": "(kW)",
+        "GCHP": "(kW)",
+        "Battery-Storage": "(kWh)",
+        "Heat Transformer": "(kW)",
+        "Electric Heating": "(kW)",
+    }
+
+    decentral_components = pd.DataFrame(columns=decentral_columns)
+    decentral_components = decentral_components.append(
+        pd.Series(decentral_columns_units), ignore_index=True
+    )
+    # creating a list to reduce the number of rows
+    decentral_components_list = [
+        "_1_pv",
+        "_2_pv",
+        "_3_pv",
+        "_4_pv",
+        "_5_pv",
+        "_gasheating_transformer",
+        "_ashp_transformer",
+        "_gchp_transformer",
+        "_battery_storage",
+        "_district_heat_link",
+        "_electricheating_transformer",
+        "_1_solarthermal_source_collector",
+        "_2_solarthermal_source_collector",
+        "_3_solarthermal_source_collector",
+        "_4_solarthermal_source_collector",
+        "_5_solarthermal_source_collector",
+    ]
+    decentral_components_from_csv = []
+    for comp in components_csv_data["ID"]:
+        i = comp.split("_")
+        if "pv" not in i[0]:
+            if "central" not in i[0]:
+                decentral_components_from_csv.append(i[0])
+    decentral_components_from_csv = set(decentral_components_from_csv)
+    # import investment values from components.csv
+    for i in decentral_components_from_csv:
+        installed_power = []
+
+        for comp in decentral_components_list:
+            # investment values of pv
[…]

[…]
