+++ program_files/processing/optimize_model.py
@@ -1,236 +1,191 @@
-"""
-    Gregor Becker - gregor.becker@fh-muenster.de
-    Christian Klemm - christian.klemm@fh-muenster.de
-"""
-
 # -*- coding: utf-8 -*-
 from oemof import solph
 from memory_profiler import memory_usage
+from datetime import datetime
 
 
 def constraint_optimization_against_two_values(
-    om: solph.Model, limit: float, storages=False
+    om: solph.Model, limit: float
 ) -> solph.Model:
     """
         Function for optimization against two parameters
         (e.g. monetary, emissions)
 
         :param om: oemof solph model to which the constraints will be \
-            added
+                   added
         :type om: oemof.solph.Model
         :param limit: maximum value for the second parameter for the \
-            whole energysystem
+                      whole energysystem
         :type limit: int
-        :param storages: boolean indicating whether or not the energy \
-            system has a storage as an investment alternative.
-        :type storages: bool
-        
-        :return: - **om** (oemof.solph.Model) - oemof solph Model \
-            within the added constraints
+        :return: - **om** (oemof.solph.Model) - oemof solph Model within \
+                                                the constraints
     """
     import pyomo.environ as po
     from oemof.solph.plumbing import sequence
-    
-    periodical_flows = {}
-    nonconvex_flows = {}
-    variable_flows = {}
-   
-    for (inflow, outflow) in om.flows:
-        # extract all investment flows where periodical constraints apply
-        if hasattr(om.flows[inflow, outflow].investment,
-                   "periodical_constraint_costs"):
-            periodical_flows[(inflow, outflow)] = \
-                om.flows[inflow, outflow].investment
-            
-        # extract all investment flows where fix constraints apply
-        if hasattr(om.flows[inflow, outflow].investment,
-                   "fix_constraint_costs")\
-                and getattr(om.flows[inflow, outflow].investment,
-                            "nonconvex"):
-            nonconvex_flows[(inflow, outflow)] = \
-                om.flows[inflow, outflow].investment
 
-        # extract all investment flows where variable constraints apply
-        if hasattr(om.flows[inflow, outflow], "emission_factor"):
-            variable_flows[(inflow, outflow)] = om.flows[inflow, outflow]
-            
+    invest_flows = {}
     ##########################
     # PERIODICAL CONSTRAINTS #
     ##########################
+    # extract all investment flows where periodical constraints apply
+    for (i, o) in om.flows:
+        if hasattr(om.flows[i, o].investment, "periodical_constraint_costs"):
+            invest_flows[(i, o)] = om.flows[i, o].investment
+    limit_name = "invest_limit_" + "periodical_constraints"
     # Setting the equation representing the sum of the periodic
-    # emissions calculated by applied investment capacity * periodical
-    # constraint costs
+    # emissions
     setattr(
         om,
-        "invest_limit_periodical_constraints",
+        limit_name,
         po.Expression(
             expr=sum(
                 om.InvestmentFlow.invest[inflow, outflow]
-                * getattr(periodical_flows[inflow, outflow],
-                          "periodical_constraint_costs")
-                for (inflow, outflow) in periodical_flows
+                * getattr(invest_flows[inflow, outflow], "periodical_constraint_costs")
+                for (inflow, outflow) in invest_flows
             )
         ),
     )
-
     ##########################
     # NONCONVEX CONSTRAINTS  #
     ##########################
+    invest_flows2 = {}
+    # extract all investment flows where fix constraints apply
+    for (i, o) in om.flows:
+        if hasattr(om.flows[i, o].investment, "fix_constraint_costs"):
+            invest_flows2[(i, o)] = om.flows[i, o].investment
+    limit_name1 = "invest_limit_" + "nonconvex_constraints"
     # Setting the equation representing the sum of the nonconvex
[…]
