+++ program_files/preprocessing/components/Storage.py
@@ -1,235 +1,168 @@
-"""
-    Christian Klemm - christian.klemm@fh-muenster.de
-    Yannick Wittor - yw090223@fh-muenster.de
-"""
 from oemof.solph.components import GenericStorage
 from oemof.solph import Flow, Investment
-import pandas
+import pandas as pd
 import logging
 
 
 class Storages:
     """
-        Creates oemof storage objects as defined in 'nodes_data' and
-        adds them to the list of components 'nodes'.
-    
-        :param nodes_data: dictionary containing parameters of the \
-            storages to be created. The following data have to be \
-            provided:
-    
-                - label
-                - active
-                - bus
-                - storage type
-                - existing capacity
-                - min. investment capacity
-                - max.investment capacity
-                - non-convex investments
-                - fix investment costs
-                - fix investment constraint costs
-                - input/capacity ratio
-                - output/capacity ratio
-                - efficiency inflow
-                - efficiency outflow
-                - initial capacity
-                - capacity min
-                - capacity max
-                - periodical costs
-                - periodical constraint costs
-                - variable input costs
-                - variable input constraint costs
-                - variable output costs
-                - variable output constraint costs
-                - capacity loss (Generic)
-                - diameter (Stratified)
-                - temperature high (Stratified)
-                - temperature low (Stratified)
-                - U value (Stratified)
-    
-        :type nodes_data: dict
-        :param busd: dictionary containing the buses of the energy system
-        :type busd: dict
-        :param nodes: list of components created before (can be empty)
-        :type nodes: list
+    Creates oemof storage objects as defined in 'nodes_data' and
+    adds them to the list of components 'nodes'.
+
+    :param nd: dictionary containing parameters of storages
+                       to be created.The following data have to be
+                       provided:
+
+                            - 'label'
+                            - 'active'
+                            - 'bus'
+                            - 'existing capacity / (kWh)'
+                            - 'min.investment capacity / (kWh)'
+                            - 'max.investment capacity / (kWh)'
+                            - 'non-convex investments'
+                            - 'fix investment costs'
+                            - 'input/capacity ratio (invest)'
+                            - 'output/capacity ratio (invest)'
+                            - 'capacity loss'
+                            - 'efficiency inflow'
+                            - 'efficiency outflow'
+                            - 'initial capacity'
+                            - 'capacity min'
+                            - 'capacity max'
+                            - 'variable input costs'
+                            - 'variable output costs'
+
+    :type nd: dict
+    :param busd: dictionary containing the buses of the energy system
+    :type busd: dict
+    :param nodes: list of components created before(can be empty)
+    :type nodes: list
+
+    Christian Klemm - christian.klemm@fh-muenster.de
     """
 
-    def create_storage(self, storage: pandas.Series, loss_rate: pandas.Series,
-                       storage_levels: list, fixed_losses: list) -> None:
-        """
-            Within this method the previously prepared storage
-            parameters are used to create an oemof storage object with
-            the given parameters and add it to the class intern list
-            nodes which is returned to the main algorithm at the end.
-            
-            :param storage: Series containing all information for \
-                the creation of an oemof storage. At least the \
-                following key-value-pairs have to be included:
-    
-                    - label
-                    - active
-                    - bus
-                    - storage type
-                    - existing capacity
-                    - min. investment capacity
-                    - max.investment capacity
-                    - non-convex investments
-                    - fix investment costs
-                    - fix investment constraint costs
-                    - input/capacity ratio
[…]
