+++ program_files/preprocessing/components/Transformer.py
@@ -1,716 +1,548 @@
-"""
-    Christian Klemm - christian.klemm@fh-muenster.de
-    Janik Budde - Janik.Budde@fh-muenster.de
-    Yannick Wittor - yw090223@fh-muenster.de
-"""
 from oemof.solph import Flow, Investment, Transformer, Source, Bus
 from oemof.solph.components import GenericCHP
 import logging
 import numpy
-import pandas
+import pandas as pd
 import os
 
 
 class Transformers:
     """
-        Creates transformer objects as defined in 'nodes_data' and adds
-        them to the list of components 'nodes'.
-    
-        :param nodes_data: dictionary containing data from excel model \
-            definition file. The following data have to be provided:
-    
-                - label
-                - active
-                - transformer type
-                - mode
-                - input
-                - input2
-                - output
-                - output2
-                - input2 / input
-                - efficiency
-                - efficiency2
-                - variable input costs
-                - variable input constraint costs
-                - variable input costs 2
-                - variable input constraint costs 2
-                - variable output costs
-                - variable output constraint costs
-                - variable output costs 2
-                - variable output constraint costs 2
-                - existing capacity
-                - max. investment capacity
-                - min. investment capacity
-                - periodical costs
-                - periodical constraint costs
-                - non-convex investment
-                - fix investment costs
-                - fix investment constraint costs
-                - heat source
-                - temperature high
-                - temperature low
-                - quality grade
-                - area
-                - length of geoth. probe
-                - heat extraction
-                - min. borehole area
-                - temp. threshold icing
-                - factor icing
-                - name
-                - electrical input conversion factor
-                - recooling temperature difference
-                - heat capacity of source
-                - min. share of flue gas loss
-                - max. share of flue gas loss
-                - min. electric power
-                - max. electric power
-                - min. electric efficiency
-                - max. electric efficiency
-                - minimal thermal output power
-                - elec. power loss index
-                - back pressure
-                
-        :type nodes_data: dic
-        :param busd: dictionary containing the buses of the energy \
-            system
-        :type busd: dict
-        :param nodes: list of components created before (can be empty)
-        :type nodes: list
+    Creates a transformer object.
+    Creates transformers objects as defined in 'nodes_data' and adds
+    them to the list of components 'nodes'.
+
+    :param nd: dictionary containing data from excel scenario
+                       file. The following data have to be provided:
+
+                            - label,
+                            - active,
+                            - transformer type,
+                            - input,
+                            - output,
+                            - output2,
+                            - efficiency,
+                            - efficiency2,
+                            - variable input costs,
+                            - variable output costs,
+                            - existing capacity,
+                            - max. investment capacity,
+                            - min. investment capacity,
+                            - periodical costs
+    :type nd: dict
+    :param busd: dictionary containing the buses of the energy system
+    :type busd: dict
+    :param nodes: list of components created before(can be empty)
+    :type nodes: list
+
+    Christian Klemm - christian.klemm@fh-muenster.de
     """
-    
+
     # intern variables
     nodes_transformer = []
     busd = None
-    
-    def get_primary_output_data(self, transformer: pandas.Series,
-                                max_invest=None) -> dict:
-        """
-            Within this method the output flow data of the currently
-            considered transformer is parameterized and returned to the
-            transformer creation method. This method results from the
-            equality of output parametrization of nearly each
-            transformer.
-            
-            :param transformer: Series containing the transformer's \
-                parameter. Series has to contain at least the \
-                following key-value pairs:
-                
-                    - min. investment capacity
[…]
