+++ program_files/postprocessing/merge_partial_results.py
@@ -1,0 +1,321 @@
+import pandas as pd
+import numpy as np
+import xlsxwriter
+import os
+from datetime import datetime
+from program_files.preprocessing.Spreadsheet_Energy_System_Model_Generator import (
+    sesmg_main, sesmg_main_including_premodel
+)
+
+
+def __get_cb_state(checkbox) -> bool:
+    """
+    removes tkinter variable handling
+
+    :param checkbox: checkbox whose boolean is to be returned
+    :type checkbox: tkinter.Variable
+    :rtype: bool
+
+    """
+    if checkbox.get() == 1:
+        return True
+    else:
+        return False
+
+
+def merge_component_csvs(limits, files, directory, result_folders):
+    components_csvs = {}
+    returns = {}
+    result_df = pd.DataFrame()
+    limits.append("1")
+    for limit in limits:
+        components_csvs.update({limit: []})
+        for scenario in result_folders[limit]:
+            components_csvs[limit].append(
+                    pd.read_csv(directory + scenario[7:] + "/components.csv")
+            )
+    for limit in limits:
+        for dataframe in components_csvs[limit]:
+            dataframe = dataframe.replace(to_replace="---", value=0)
+            result_df = result_df.append(dataframe, sort=True)
+        for col in result_df.keys():
+            if not (str(col) == "ID" or str(col) == "type"):
+                result_df[str(col)] = result_df[str(col)].astype(float)
+        result_df = result_df.groupby(["ID", "type"]).sum(axis=0)
+        returns.update({limit: result_df})
+    for result in returns:
+        returns[result].to_csv(directory + "/components_{}.csv".format(result))
+    return returns
+
+
+def calc_constraint_limits(result_folders, limits):
+    constraints = {}
+    # get constraints of the first optimization
+    result = pd.read_csv(str(result_folders["1"][0]) + "/components.csv")
+    constr_min_1 = float(sum(result["constraints/CU"]))
+    # get constraints of the second optimization
+    result2 = pd.read_csv(str(result_folders["0"][0]) + "/components.csv")
+    constr_min_2 = float(sum(result2["variable costs/CU"])
+                         + sum(result2["periodical costs/CU"]))
+    # devide solvable range in "limits" intervals
+    for i in limits:
+        constraints.update({i: constr_min_1 - float(
+            constr_min_1 - constr_min_2) * float(i)})
+    return constraints
+
+
+def criterion_switch_dh(directory):
+    columns = {}
+    path = os.path.dirname(os.path.dirname(directory)) \
+           + "/program_files/urban_district_upscaling/standard_parameters.xlsx"
+    # get keys from plain scenario
+    standard_parameter = pd.ExcelFile(path)
+    # get columns from plain sheet
+    for sheet in standard_parameter.sheet_names:
+        if sheet not in ["8_1_other"]:
+            columns[sheet] = standard_parameter.parse(sheet)
+    
+    other_param = standard_parameter.parse("8_1_other", index_col="label")
+    writer = pd.ExcelWriter(path, engine="xlsxwriter")
+    
+    costs = other_param.loc[:, "costs"].copy()
+    other_param.loc[:, "costs"] = \
+        other_param.loc[:, "constraint costs"]
+    other_param.loc[:, "constraint costs"] = costs
+    other_param.reset_index(inplace=True, drop=False)
+    other_param.to_excel(writer, "8_1_other", index=False)
+    for sheet in columns:
+        columns[sheet].to_excel(writer, sheet, index=False)
+    writer.save()
+
+
+def run_pareto(
+        limits, scenario, gui_variables, timeseries_prep_param, pre_modeling,
+        pre_model_timeseries_prep, investment_boundaries,
[…]
