+++ program_files/GUI_files/urban_district_upscaling_GUI.py
@@ -1,0 +1,159 @@
+import os
+from program_files import urban_district_upscaling_pre_processing
+from program_files.urban_district_upscaling.urban_district_upscaling_post_processing import (
+    urban_district_upscaling_post_processing,
+)
+from program_files.urban_district_upscaling.urban_district_upscaling_post_processing_clustered import (
+    urban_district_upscaling_post_processing_clustered,
+)
+import subprocess
+
+
+class UpscalingFrameClass:
+    """
+    This class is used to create the Graphical User Interface
+    (GUI) for the Urban_District_Upscaling_Tool.
+    In this context, it uses the methods of the superclass
+    MethodsGUI.
+
+    :param frame: tkinter upscaling frame
+    :type frame: ttk.Frame
+    :param gui_variables: dictionary containing GUI variables
+    :type gui_variables: dict
+    :param tk: object containing the super class methods
+    :type tk: GUI object
+
+    """
+
+    @staticmethod
+    def scenario_upscaling(
+        pre_scenario, standard_param, scenario_name, clustering, clustering_dh
+    ):
+        """
+        Methods starting the upscaling pre_processing Algorithm
+
+        :param pre_scenario: containing path to pre_scenario file
+        :type pre_scenario: tk.StringVar
+        :param standard_param: containing path to standard_parameter
+            file
+        :type standard_param: tk.StringVar
+        :param scenario_name: containing path to scenario_name file
+        :type scenario_name: tk.StringVar
+        :param clustering: containing boolean rather the pre
+            scenario is clustered or not
+        :type clustering: tk.BooleanVar
+        :param clustering_dh:
+        :type clustering_dh:
+        """
+        urban_district_upscaling_pre_processing(
+            paths=[
+                pre_scenario.get(),
+                standard_param.get(),
+                scenario_name.get(),
+                os.path.join(
+                    os.path.dirname(__file__),
+                    r"../urban_district_upscaling/plain_scenario.xlsx",
+                ),
+            ],
+            clustering=clustering.get(),
+            clustering_dh=clustering_dh.get(),
+        )
+
+    @staticmethod
+    def create_overview(components, clustering):
+        """
+        Methods starting the upscaling post_processing Algorithm
+
+        :param components: containing path to components.csv
+        :type components: tk.StringVar
+        :param clustering: containing boolean rather the pre
+            scenario is clustered or not
+        :type clustering: tk.BooleanVar
+
+        """
+        if clustering:
+            urban_district_upscaling_post_processing_clustered(components.get())
+        else:
+            urban_district_upscaling_post_processing(components.get())
+        subprocess.call(os.path.dirname(__file__) + "/overview.xlsx", shell=True)
+
+    def __init__(self, frame, gui_variables, tk):
+        # Headline
+        row = 0
+        tk.create_heading(
+            frame, "Urban District Upscaling", 0, row, "w", True, columnspan=7
+        )
+        # Description
+        row += 1
+        tk.create_heading(
+            frame,
+            "Standardized implementation of " "urban energy systems",
+            0,
+            row,
+            "w",
+            columnspan=7,
+        )
+        row += 1
+        # Headline
+        tk.create_heading(frame, "Preprocessing", 0, row, "w", True)
+        row += 1
+        upscaling_elements = {
+            "Pre-Scenario": [
+                lambda: tk.get_path("xlsx", gui_variables["pre_scenario_path"]),
+                "Change",
+                "pre_scenario_path",
+            ],
+            "Standard Parameters": [
+                lambda: tk.get_path("xlsx", gui_variables["standard_parameter_path"]),
+                "Change",
+                "standard_parameter_path",
+            ],
+            "Scenario Name": [
+                lambda: tk.get_path("xlsx", gui_variables["scenario_name"]),
[…]

[…]
