+++ program_files/GUI_st/GUI_old_comment.py
@@ -1,0 +1,552 @@
+# -*- coding: utf-8 -*-
+from datetime import datetime
+from tkinter import *
+from tkinter import ttk
+import subprocess
+import os
+import csv
+from program_files.preprocessing.Spreadsheet_Energy_System_Model_Generator \
+    import \
+    sesmg_main
+from program_files.GUI_files.urban_district_upscaling_GUI \
+    import UpscalingFrameClass
+from program_files.GUI_files.MethodsGUI import MethodsGUI
+import \
+    program_files.postprocessing.merge_partial_results as merge_partial_results
+import program_files.postprocessing.plotting as plotting
+from program_files.preprocessing.create_energy_system import import_scenario
+
+
+def get_pid():
+    """ Returns the ID of the running process on Port 8050 """
+    # TODO @JAN diese Methode war notwendig, um den alten plotly dash Prozess
+    # über seine IP und seinen Port zu clearen, und anschließend einen neuen
+    # zu starten. Ich denke die benötigen wir in der neuen Struktur nicht mehr
+    import socket
+    import errno
+    import sys
+    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    try:
+        # Checks if port 8050 can be reached
+        s.bind(("127.0.0.1", 8050))
+        # If Yes, the program continues in line 53
+        closeProcess = False
+    # If not, the ID of the running process is returned
+    except socket.error as e:
+        if e.errno == errno.EADDRINUSE:
+            closeProcess = True
+        else:
+            raise ValueError("closeProcess failure")
+    s.close()
+    if closeProcess:
+        if sys.platform.startswith("win"):
+            command = "netstat -aon| findstr :8050"
+        elif sys.platform.startswith("darwin"):
+            command = "lsof -i tcp:8050"
+        elif sys.platform.startswith("linux"):
+            command = "fuser -n tcp 8050"
+        else:
+            raise ValueError("unknown platform")
+        # list of processes running on 8050
+        p_ids_list = str(subprocess.check_output(command, shell=True)).split()
+        if sys.platform.startswith("win"):
+            pid = p_ids_list[5]
+            pid = pid[:-4]
+        elif sys.platform.startswith("darwin"):
+            pid = p_ids_list[9]
+        elif sys.platform.startswith("linux"):
+            pid = p_ids_list[1]
+        else:
+            raise ValueError("unknown platform")
+        return pid
+    else:
+        return ''
+
+
+def save_settings(gui_variables: dict):
+    """
+        Method that stores the settings entered in the GUI in a csv file
+        in program_files/technical_data.
+
+        :param gui_variables: dictionary holding the settings chosen in
+            GUI
+        :type gui_variables: dict
+
+    """
+    # TODO @JAN diese Methode speichert bei einem gestarteten Lauf alle
+    # in der GUI getroffenen Einstellungen in einer csv, die in der Ordner-
+    # struktur des SESMGs abgelegt wird. Hier haben wir gestern besprochen,
+    # dass dies vielleicht erst der zweite Schritt bei der Entwicklung der
+    # GUI sein sollte.
+    dict_to_save = {}
+    # remove variable handling from tkinter
+    for key, value in gui_variables.items():
+        dict_to_save.update({key: value.get()})
+    # store extracted variables to csv file
+    with open('program_files/technical_data/gui_settings.csv', 'w',
+              newline='') as csv_file:
+        writer = csv.writer(csv_file)
+        for key, value in dict_to_save.items():
+            writer.writerow([key, value])
+
+
+def reload_settings(gui_variables: dict):
+    """
+        Method that loads the settings saved in the csv file and then
+        embeds them in the GUI.
+
[…]
