+++ program_files/Demo_Tool/demo_tool.py
@@ -1,0 +1,497 @@
+from tkinter import *
+import tkinter as ttk
+import pandas as pd
+import openpyxl
+import matplotlib as mlp
+import matplotlib.pyplot as plt
+import numpy as np
+from threading import *
+import os
+import subprocess
+from program_files.preprocessing.Spreadsheet_Energy_System_Model_Generator import (
+    sesmg_main,
+)
+from program_files.GUI_files import MethodsGUI
+
+
+class demo_frame_class(MethodsGUI):
+    def execute_sesmg_DEMO(self, demo_file, demo_results):
+        """Excecutes the optimization algorithm"""
+        print(demo_file)
+        print(demo_results)
+
+        sesmg_main(
+            scenario_file=demo_file,
+            result_path=demo_results,
+            num_threads=2,
+            timeseries_prep=["none", "none", "none", "none", 0],
+            graph=False,
+            criterion_switch=False,
+            xlsx_results=False,
+            console_results=False,
+            solver="cbc",
+            cluster_dh=False,
+            district_heating_path="",
+        )
+
+    def demo_scenario(self, mode):
+        """modifies financial demo scenario"""
+        mode_dict = {
+            "monetary": [
+                "demo_scenario_monetaer.xlsx",
+                r"/results/demo/financial",
+                "Total System Costs",
+                "Total Constraint Costs",
+            ],
+            "emissions": [
+                "demo_scenario_emissionen.xlsx",
+                r"/results/demo/emissions",
+                "Total Constraint Costs",
+                "Total System Costs",
+            ],
+        }
+
+        xfile = openpyxl.load_workbook(
+            os.path.dirname(__file__)
+            + "/v0.0.6_demo_scenario/"
+            + mode_dict.get(mode)[0],
+            data_only=True,
+        )
+
+        # PHOTOVOLTAICS
+        sheet = xfile["sources"]
+        sheet["I3"] = int(self.entry_values["photovoltaics"].get())
+        sheet["J3"] = int(self.entry_values["photovoltaics"].get())
+        # SOLAR THERMAL
+        sheet = xfile["sources"]
+        sheet["I5"] = int(self.entry_values["solarthermal"].get())
+        sheet["J5"] = int(self.entry_values["solarthermal"].get())
+        # BATTERY
+        sheet = xfile["storages"]
+        sheet["N3"] = int(self.entry_values["battery"].get())
+        sheet["O3"] = int(self.entry_values["battery"].get())
+        # CHP
+        sheet = xfile["transformers"]
+        sheet["C4"] = int(self.entry_values["district heating"].get())
+        sheet["L4"] = int(self.entry_values["chp"].get())
+        sheet["M4"] = int(self.entry_values["chp"].get())
+        # ASHP
+        sheet = xfile["transformers"]
+        sheet["L6"] = int(self.entry_values["ASHP"].get())
+        sheet["M6"] = int(self.entry_values["ASHP"].get())
+        # GCHP
+        sheet = xfile["transformers"]
+        sheet["L5"] = int(self.entry_values["GCHP"].get())
+        sheet["M5"] = int(self.entry_values["GCHP"].get())
+        # THERMAL STORAGE
+        sheet = xfile["storages"]
+        sheet["C4"] = int(self.entry_values["district heating"].get())
+        sheet["N4"] = int(self.entry_values["thermal storage"].get())
+        sheet["O4"] = int(self.entry_values["thermal storage"].get())
+        # THERMAL STORAGE
+        sheet = xfile["storages"]
+        sheet["N5"] = int(self.entry_values["thermal storage (decentralized)"].get())
+        sheet["O5"] = int(self.entry_values["thermal storage (decentralized)"].get())
+        # District Heating
[…]
