+++ program_files/preprocessing/pre_model_analysis_20230113.py
@@ -1,252 +1,0 @@
-import pandas as pd
-import shutil
-
-def filter_result_component_types(components, type):
-    ' returns dataframe containing only one specific component type'
-
-    filtered_components = components[(components.type == type)]
-    return(filtered_components)
-
-def update_component_investment_decisions(pre_model_results, scenario_path, results_components_path, updated_scenario_path, scenario_type_name, result_type_name, boundary_factor, investment_boundaries=True):
-    '''adapts investment decision depending on the results of a pre-model and returns new dataset plus list of
-    deactivated components'''
-    components = pre_model_results
-    # return components type results
-    result_components = filter_result_component_types(components=components, type=result_type_name)
-
-    # return component type scenario sheet
-    components_xlsx = pd.read_excel(scenario_path, sheet_name=scenario_type_name)
-    # drop first (nan) column
-    components_xlsx = components_xlsx.iloc[1: , :]
-    # reset of index required, so that it is uniform to the result-dataframe
-    components_xlsx = components_xlsx.reset_index()
-
-    if scenario_type_name == 'district heating':
-        dh_technical_pre_selection(components_xlsx, result_components)
-        list_of_deactivated_components = []
-
-        return components_xlsx, list_of_deactivated_components
-
-    elif scenario_type_name == 'buses':
-        bus_technical_pre_selection(components_xlsx, result_components)
-        list_of_deactivated_components = []
-
-        return components_xlsx, list_of_deactivated_components
-
-    elif result_type_name == 'insulation':
-        insulation_technical_pre_selection(components_xlsx, result_components)
-        list_of_deactivated_components = []
-
-        return components_xlsx, list_of_deactivated_components
-
-    else:
-        list_of_deactivated_components = technical_pre_selection(components_xlsx, result_components)
-        if investment_boundaries == True:
-            tightening_investment_boundaries(components_xlsx, result_components, boundary_factor)
-
-        return components_xlsx, list_of_deactivated_components
-
-def technical_pre_selection(components_xlsx, result_components):
-    '''deactivates investment-components for which no investments has been carried out and additionally returns a list
-    of deactivated components'''
-    list_of_deactivated_components = []
-    result_components.set_index('ID', inplace=True)
-    for i,scenario_component in components_xlsx.iterrows():
-        if scenario_component['label'] in result_components.index.values:
-            if float(result_components.loc[scenario_component['label']]['max. invest./kW']) > 0:
-                if str(result_components.loc[scenario_component['label']]['investment/kW']) in ['0.0','0','0.00','---','-0.0', '-0.00']:
-                    components_xlsx.at[i, 'active'] = 0
-                    list_of_deactivated_components.append(str(scenario_component['label']))
-
-
-
-        # if float(result_components.iloc[i]['max. invest./kW']) > 0:
-        #     if str(result_components.iloc[i]['investment/kW']) == '0.0':
-        #         components_xlsx.at[i, 'active'] = 0
-        #         list_of_deactivated_components.append(str(scenario_component['label']))
-
-    return list_of_deactivated_components
-
-    deactive_respective_competition_constraints(scenario_path, list_of_deactivated_components)
-
-def tightening_investment_boundaries(components_xlsx, result_components, boundary_factor):
-    'tightens investment boundaries'
-    list_of_adapted_components = []
-    for i,scenario_component in components_xlsx.iterrows():
-        if scenario_component['label'] in result_components.index.values:
-            if float(result_components.loc[scenario_component['label']]['max. invest./kW']) > 0:
[…]

[…]
