### 2023-06-19 f984674
installation( ): reset installation to the TK inter installation
