+++ Linux_installation.sh
@@ -33,7 +33,6 @@ sudo mv Run_SESMG_for_windows.cmd files_deviant_operating_system
 echo ----------------------
 echo Installation completed
 echo ----------------------
-python3 -m streamlit run "program_files/GUI_st/1_Main_Application.py"
 
 exit
 
