+++ Run_SESMG_for_Linux.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 . bin/activate
-python3 -m streamlit run "program_files/GUI_st/1_Main_Application.py"
-
+python3 start_script.py
