### 2023-06-19 cc70fa5
installation( ): add basemap
