+++ requirements.txt
@@ -1,8 +1,9 @@
 streamlit>=1.17.0
 streamlit-aggrid>=0.3.0
 oemof.solph==0.4.5
+dhnx @ Https://github.com/SESMG/DHNx/archive/refs/heads/dev.zip
 openpyxl>=3.0.0  
-oemof.thermal>=0.0.5  
+oemof.thermal==0.0.5  
 feedinlib @ https://github.com/SESMG/feedinlib/archive/refs/heads/dev.zip
 open_fred-cli>=0.0.1
 geocoder>=1.38.0
@@ -12,11 +13,11 @@ richardsonpy>=0.2.0
 graphviz>=0.20  
 scikit-learn-extra>=0.2.0  
 memory-profiler>=0.60.0  
-dhnx @ Https://github.com/SESMG/DHNx/archive/refs/heads/dev.zip
 sympy>=1.10.0   
 xlsxwriter>=3.0.0  
 seaborn>=0.11.0  
 dash>=2.4.0  
 dash_canvas>=0.1.0 
+basemap==1.3.7
 h5py==3.6.0
 pytest
