### 2023-06-20 SESMG/SESMG#188 @willu47
[JOSS] Installation
Hi, this is a question regarding installtion from [JOSS Review #5519](https://github.com/openjournals/joss-reviews/issues/5519)

Please could you explain to me why you have decided against using an automated package management solution such as setuptools, conda or venv to handle the installation of dependencies for SESMG?

As a MacOS + conda user, I was unable to follow the installation instructions, as they make the assumption that I am installing Python from scratch on a systemwide basis.

I installed the package easily by using the following commands:

    git clone https://github.com/SESMG/SESMG.git
    conda create -n sesmg python=3.10 coincbc geos graphviz postgresql
    pip install -r requirements.txt

    mkdir files_deviant_operating_system
    mv Linux_installation.sh files_deviant_operating_system
    mv Run_SESMG_for_Linux.sh files_deviant_operating_system
    mv Windows_installation.cmd files_deviant_operating_system
    mv Run_SESMG_for_windows.cmd files_deviant_operating_system

    python -m streamlit run "program_files/GUI_st/1_Main_Application.py" osascript -e 'tell application "Terminal" to quit'

But I don't understand why you have you not packaged the SESMG as a Python package.  There's nothing in the above which could not be solved by a Python package, allowing users to `pip install sesmg`.  You could then install a command line hook to run streamlit e.g. `sesmg run`.  Using setuptools you can detect the environment, if you need to call custom setup scripts.

In other words, your current installation instructions are too opinionated. They force users into a way of working which is not very flexible.  I would recommend that you split your installation instructions into two parts 1. installation of dependencies (CBC, postgres, geos, graphviz) and provide alternative means to do this e.g. conda, brew, manual download + install and 2. pip install sesmg
