### 2023-06-20 SESMG/SESMG#188 @chrklemm
Dear @willu47,
thank you for your effort and your feedback.

Within the automatic installation, which is executed with the respective "installation" file, a virtual environment is automatically set up. We decided to use an installation procedure that makes the SESMG usable without any programming knowledge. This means that no advanced IDEs are required and we do not use terminal commands. We hope that this will provide a low-threshold entry, so that users without programming and modeling knowledge can also use the SESMG.

However, we will discuss your recommendation in the dev-team and look for a solution that is convenient for all types of users.

By the way: We saw that you used Python 3.10, which support is currently not guaranteed. We recommend to use Python 3.7, 3.8 or 3.9 (see https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#macos ).

Best regards,
SESMG-Team
