### 2023-06-26 SESMG/SESMG#190 @nick-gorman
[JOSS] Unable to run demo tool
**Code of Conduct**
- [x] Hereby I consent to the [CODE_OF_CONDUCT](https://github.com/SESMG/SESMG/blob/master/CODE_OF_CONDUCT.md)

**Describe the bug**
A clear and concise description of what the bug is.

**To Reproduce**
Steps to reproduce the behavior:
1. Go to Demo Tool
2. Change Ground coupled heat pump to 24 kW
3. Change District Heating Network to Urban
4. Click Start Simulation

**Expected behavior**
Not sure what should happen, but nothing does. In the cmd window an error appears, see below.

```
INFO:root:Cartopy is not installed. Background maps will not be drawn.
20:48:32-INFO-Path for logging: C:\Users\nick\Documents\GitHub\SESMG\results\demo\oemof.log
20:48:35-INFO-   Spreadsheet scenario successfully imported.
20:48:35-INFO-   Start import weather data
2023-06-26 20:48:35.118 Uncaught app exception
Traceback (most recent call last):
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\streamlit\runtime\scriptrunner\script_runner.py", line 552, in _run_script
    exec(code, module.__dict__)
  File "C:\Users\nick\Documents\GitHub\SESMG\program_files\GUI_st\pages\4_Demo_Tool.py", line 365, in <module>
    create_demo_model_definition()
  File "C:\Users\nick\Documents\GitHub\SESMG\program_files\GUI_st\pages\4_Demo_Tool.py", line 258, in create_demo_model_definition
    execute_sesmg_demo(
  File "C:\Users\nick\Documents\GitHub\SESMG\program_files\GUI_st\pages\4_Demo_Tool.py", line 181, in execute_sesmg_demo
    sesmg_main(
  File "C:\Users\nick\Documents\GitHub\SESMG\program_files\preprocessing\Spreadsheet_Energy_System_Model_Generator.py", line 77, in sesmg_main
    nodes_data = create_energy_system.import_model_definition(
  File "C:\Users\nick\Documents\GitHub\SESMG\program_files\preprocessing\create_energy_system.py", line 82, in import_model_definition
    nodes_data = import_open_fred_weather_data(nodes_data, lat, lon)
  File "C:\Users\nick\Documents\GitHub\SESMG\program_files\preprocessing\import_weather_data.py", line 71, in import_open_fred_weather_data
    geo_info = geocoder.osm([lat, lon], method='reverse')
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\geocoder\api.py", line 458, in osm
    return get(location, provider='osm', **kwargs)
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\geocoder\api.py", line 198, in get
    return options[provider][method](location, **kwargs)
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\geocoder\base.py", line 395, in __init__
    self.params = OrderedDict(self._build_params(location, provider_key, **kwargs))
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\geocoder\osm_reverse.py", line 25, in _build_params
    'q': str(Location(location)),
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\geocoder\location.py", line 22, in __init__
    self._check_input(location)
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\geocoder\location.py", line 53, in _check_input
    self._check_for_list(location)
  File "C:\Users\nick\Documents\GitHub\SESMG\lib\site-packages\geocoder\location.py", line 88, in _check_for_list
    raise ValueError(
ValueError: Coords are not within the world's geographical boundary
```

**Screenshots**
![image](https://github.com/SESMG/SESMG/assets/40549624/6601fb4b-4d5c-48fb-b875-7aa756751951)

 - OS: Windows
 - Browser: Chrome
 - Version: v0.5.0

**Additional context**
When I installed SESMG I also got the error message described in #189
