### 2023-06-26 329e3d2
docu(trouble_shooting.rst): improve documentation about #189
