### 2023-06-26 SESMG/SESMG#190 @chrklemm
Dear @nick-gorman,

thank you very much for your contribution. We have tried to reproduce this error. We only managed to do this by modifying the following file: "SESMG/program_files/demo_tool/v0.4.0_demo_model_definition/demo_model_definition.xlsx"

The demo tool and the associated model definition are a self-contained system and it is not intended to modify the related files. Changes in the demo tool are made by the user through the GUI only. All other models, which are started on the "main page" of the gui, for example "SESMG/model_definition_example.xlsx" may be adapted.

Can you provide the following information so that we can follow up on the issue?
- Which Python version are you using?
- Are you using the SESMG version from the main branch?
- Have you modified this file ("SESMG/program_files/demo_tool/v0.4.0_demo_model_definition/demo_model_definition.xlsx")?
- Did you manually change or reinstall packages, e.g. to solve Issue #189?

Thank you in advance and best regards,
SESMG-Dev-Team
