### 2023-06-26 98b94fd
docu(trouble_shooting.rst): fix docu design
