### 2023-06-26 7da178a
docu(installation.rst): update installation instructions about #189
