### 2023-06-26 8694984
docu(installation.rst): fix link
