### 2023-06-26 d011adf
docu(paper.md): typo
