### 2023-06-27 SESMG/SESMG#190 @nick-gorman
@chrklemm

- I'm using python 3.8
- Using master branch as of commit: 27b0e8e177c3e75bdde19df3ec6c457f0fa04d51
- No, I have not modified the file "SESMG/program_files/demo_tool/v0.4.0_demo_model_definition/demo_model_definition.xlsx"
- No didn't do any manual install process, just cloned the repo, installed graphviz, got the cbc binaries, and ran the install cmd file

Would appreciate any further suggestions
