### 2023-06-27 SESMG/SESMG#190 @jtock
Hi @nick-gorman,

we figured out that there is an incompatibility between some versions of python 3.8 and some sub packages we are using. E.g. we were able to reproduce your issue with the python version 3.8.4 to 3.8.7, due to an incompatibility issue with pandas during the import of xlsx-files.

The installation and program is running normally with the python versions we are usually using, which are e.g. 3.8.10 or 3.9.13.

We are working on a way to automatically test the Python subversions, to determine if there are more incompatibilities. If you want to continue your review you could use the most extensively tested version 3.9.13, till we come up with a satisfactory solution.

Greeting
SESMG Dev Team
