### 2023-06-28 5fbdd90
docu(installation): update regarding #188
