### 2023-06-28 6d61716
docu(installation): adaptions regarding issue #188
