### 2023-06-28 SESMG/SESMG#190 @GregorBecker
Hello @nick-gorman,

we have now been able to reproduce your error, and have discovered that it is caused by the new release of pandas 2.0., which is automatically  installed in some Python versions. As a result, we have now removed this version from the suitable requirements (pandas < 2.0.0), ensuring that functionality is restored.

Your issue should be solved by installing the SESMG again from scratch. If this is not the case, please contact us again.

Kind regards
SESMG Dev Team
