### 2023-06-28 SESMG/SESMG#188 @GregorBecker
Hello @willu45,

first of all we would like to thank you again for your comments regarding our installation process. We agree with you, that it was very specific and therefore did not allow the use of the SESMG in a different virtual environment than the one we specified.

After discussing and trying to implement the packaging e.g. via PyPI in the team, we figured our that this kind of deployment is not suitable for the SESMG, due to the following reasons:

- The SESMG is not a typical python package and is therefore not imported in the classic sense, but rather represents an application, which should therefore be seen as a closed product.

- The SESMG needs the given folder structure on Github. During the optimization process the results are stored in the result folder, for example. Those paths are fixed in the code and cannot be influenced by the user.

- PyPI only provides an installation of sub-packages which are also shared via PyPI. We are using some packages which are not shared this way, which is a third barrier to use PyPI.

After we have looked at the packaging and have determined that a PyPI packaging is not purposeful, we have decided to extend the installation description by an advanced user entry, which allows the user to use other environments (e.g. Anaconda) as shown in your code section. We updated the  documentation about this additional way to install the SESMG, which you can find here: https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#advanced

We hope to aim your ideas of an alternative installation process. Otherwise, please feel free to contact us again.

Kind regards
SESMG Dev Team
