### 2023-06-30 209e015
docu(trouble_shooting.rst): add python version error from #191
