+++ docs/03.00.00_trouble_shooting.rst
@@ -75,17 +75,21 @@ Error I-006: Could not find a local HDF5 installation (macOS M1/M2)
 
 **Debugging:** Type following lines separately in the terminal: ``(echo; echo 'export HDF5_DIR=/opt/homebrew/opt/hdf5') >> /Users/$(whoami)/.zprofile`` and ``(echo; echo 'export BLOSC_DIR=/opt/homebrew/opt/c-blosc') >> /Users/$(whoami)/.zprofile``.
 
-Error I-007: No module named streamlit
+Error I-007: No module named streamlit I
 -------------------------------------------------------------------
 **Error Message:** ::
 
    No module named streamlit
 
-**Possible Error Cause:** This issue can occur at the first model run after installation, even if there was no error message. Maybe you typed in the wrong python version at the installation.  
+**Possible Error Cause I:** This issue can occur at the first model run after installation, even if there was no error message. Maybe you typed in the wrong python version at the installation.  
 
-**Debugging:** You need to delete the folder and restart the installation with a 'freshly' unpacked .zip file (downloadable on Github). Make shure to type your installed python version in the major.minor format, which means to give 3.X if you have the version 3.X.YY.
+**Debugging I:** You need to delete the folder and restart the installation with a 'freshly' unpacked .zip file (downloadable on Github). Make shure to type your installed python version in the major.minor format, which means to give 3.X if you have the version 3.X.YY.
 
-Error I-008: No module named streamlit
+**Possible Error Cause II:** The SESMG is compatible with a specific set of Python versions. The problem can occure if you are ursing an unsupported version. You can find a list of the supported Python versions in the  `installation instructions <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#installation>`_.   
+
+**Debugging II:** You can install an additional Python version on your machine which is facing the SESMGs requirements. Afterwards you need to restart the installation process from scratch in a new folder.
+
+Error I-008: No module named streamlit II
 -------------------------------------------------------------------
 **Error Message:** ::
 
