+++ docs/02.02.03_urban_district_upscaling.rst
@@ -83,12 +83,12 @@ Category 2
 
 
 .. csv-table:: Input for the upscaling sheet. Category 2: Building investment data.
-   :header:  label,HS,ashp,gchp,parcel ID,oil heating,gas heating,battery storage,thermal storage,central heat,electric heating,st or pv 1,roof area 1,surface tilt 1,azimuth 1,st or pv 2,roof area 2,surface tilt 2,azimuth 2
+   :header:  label,HS,ashp,gchp,parcel ID,oil heating,gas heating,battery storage,thermal storage,central heat,electric heating,wood stove,aahp,st 1,pv 1,roof area 1,surface tilt 1,azimuth 1,st 2,pv 2,roof area 2,surface tilt 2,azimuth 2
 
-   x,,,,,,,,,,,,(m²),(°),(°),,(m²),(°),(°)
-   001_building,1,no,no,no,no,no,no,no,yes,no,0,0,0,0,0,0,0,0
-   002_building,1,no,no,no,no,yes,no,no,no,no,pv&st,150,75,100,0,0,0,0
-   003_building,1,yes,yes,GCHP25,no,no,yes,yes,yes,no,pv&st,200,50,180,0,0,0,0
+   x,,,,,,,,,,,,,,,(m²),(°),(°),,,(m²),(°),(°)
+   001_building,1,no,no,no,no,no,no,no,yes,no,yes,yes,no,no,0,0,0,no,no,0,0,0
+   002_building,1,no,no,no,no,yes,no,no,no,no,yes,yes,yes,yes,150,75,100,0,0,0,0,0
+   003_building,1,yes,yes,GCHP25,no,no,yes,yes,yes,no,yes,yes,yes,yes,200,50,180,0,0,0,0,0
 
 
 - **label**: The building name can be chosen by the user and is the identification number (ID) of a building. The ID must be unique for each building, because all the following columns are assigned to it.
@@ -96,11 +96,12 @@ Category 2
 - **gchp**: Ground-coupled heat pumps are limited by the area required for geothermal collectors or probes. If there is a potential area for the GCHP, the so-called parcel must be assigned to the buildings.
 - **parcel ID**: The parcel ID assigns a potential area for GCHP to the buildings. On an additional auxiliary data sheet, users enter the parcel ID and the potential area.
 - **heat extraction (kW/m)**: The extraction capacity of the geothermal probes or collectors is crucial for the performance of the heat pumps. The extraction rate should be determined specifically for the location.
-- **oil heating, gas heating, electric heating, battery storage, thermal storage, wood stove**: The technologies are not subject to restrictions and can be considered as an investment alternative.
+- **oil heating, gas heating, electric heating, battery storage, thermal storage, wood stove, aahp**: The technologies are not subject to restrictions and can be considered as an investment alternative.
 - **central heat**: If a heating network is available, a network connection can be considered as an investment alternative.
 - **wood stove share**: If a value between 0 and 1 is used as "wood stove share", this share is separated from the main household heat demand and connected primarily to the wood stove. To avoid insolubility, this bus is also connected to the main heating bus of the considered building. If this division of the sink is not desired, "standard" should be used.
 - **solar thermal share**: If a value between 0 and 1 is used as "solar thermal share", this share is separated from the main household heat demand and connected primarily to the solar thermal collector. To avoid insolubility, this bus is also connected to the main heating bus of the considered building. If this division of the sink is not desired, "standard" should be used.
-- **st or pv 1**: In this column it is decided whether the roof potential area applies to solar thermal (ST) systems, photovoltaic (PV) systems or to both systems. As soon as both systems are relevant for one area, an area competition arises, which is automatically considered.
+- **st 1**: In this column it is decided whether the roof potential area applies to solar thermal (ST) systems. Possible entries: yes or no.
[…]
