+++ docs/02.02.02_model_definition.rst
@@ -329,7 +329,7 @@ Compression Heat Transformer
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 The following parameters are only required, if "transformer type" is set to "CompressionHeatTransformer":
 
-- **heat source**: Specifies the heat source. Possible heat sources are "GroundWater", "Ground", "Air" and "Water" possible.
+- **heat source**: Specifies the heat source. Possible heat sources are "GroundWater", "Ground", "Air", "Air-to-Air" (which represents an AAHP) and "Water" possible.
 - **temperature high** in (°C): Temperature of the high temperature heat reservoir. Only required if "mode" is set to "heat_pump".
 - **temperature low** in (°C): Cooling temperature needed for cooling demand. Only required if "mode" is set to "chiller".
 - **quality grade**: To determine the COP of a real machine a scale-down factor (the quality grade) is applied on the Carnot efficiency (see `oemof.thermal <https://github.com/wind-python/windpowerlib/blob/dev/windpowerlib/oedb/turbine_data.csv>`_).
