### 2023-07-04 64241b0
docu(Readme): Adaptions regarding issue #188
