### 2023-07-04 b07255e
docu(readme): typo
