+++ README.md
@@ -27,7 +27,6 @@ Furthermore, the **SESMG** comes with important modeling methods, enabling holis
 ## Quick Start 
 [![Generic badge](https://img.shields.io/badge/content-how-green.svg)](https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/#)
 
-#### Please note: 
 A detailed description of the installation process for Windows, MacOS and Linux can be found in the [documentation (chapter installation)](
 https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html). 
 
