### 2023-07-04 SESMG/SESMG#188 @willu47
Hi @GregorBecker. Thanks for the responses. Given the opportunity of the review, I would like to challenge you and the team on each of these points.

> - The SESMG is not a typical Python package and is therefore not imported in the classic sense, but rather represents an application, which should therefore be seen as a closed product.

Then why not use a more complete packaging solution?  There are various solutions such as `py2exe` and `pyinstaller` that allow you to create an installable package.  You can also bundle a Flask app, freezing all dependencies and a Python version.  This would be a more complete solution to what you have now, which seems a bit in the middle.

> - The SESMG needs the given folder structure on Github. During the optimization process the results are stored in the result folder, for example. Those paths are fixed in the code and cannot be influenced by the user.

This is a design issue that could be avoided by using the Python provisions for bundling data files in applications and having access to the system and temporary folders.

See the [setuptools](https://setuptools.pypa.io/en/latest/userguide/datafiles.html) documentation for instructions on the former.

For the latter, you could allow the user to define a results folder in the GUI, or write results to a temporary directory using one of the inbuilt Python libraries (which work across filesystems). In general, you want to avoid hardcoding file paths.

> - PyPI only provides an installation of sub-packages which are also shared via PyPI. We are using some packages which are not shared this way, which is a third barrier to use PyPI.

You can use pip to install Python packages from any location on the internet, including Git repositories, other Python packaging indexes etc.  The issue is some of the more hard-to-install packages such as GDAL which can be a pain to install on Windows if you're not using conda. But then that is an argument for using something like miniconda which will also help manage dependency compatibility.

Another solution could be to use a virtual machine to distribute the program, such as a Docker container or vagrant.
---
In summary, there's no perfect way to package and deploy your Python application. I don't think the solution you have at the moment is optimal, given your focus on ease of use.
