### 2023-07-04 SESMG/SESMG#193 @willu47
[JOSS] Paper
Hi. In the context of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5519) I have a few comments on the content of the paper.

Please double check you are consistent with British/American English.

- [x] A list of the authors of the software and their affiliations, using the correct format (see the example below).
- [x] A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

  - [x] I think you could more clearly outline what functionality the software provides e.g. a browser-based graphical user interface, spreadsheets to provide data entry, comprehensive documentation on how to use it

- [ ] A Statement of Need section that clearly illustrates the research purpose of the software and places it in the context of related work.

  - [ ] This includes many technical terms that are not defined e.g. "sector coupling".
  - [ ] Suggest replacing 'volatile' with 'variable'
  - [ ] The current text in this part of the paper is more focussed on general energy modelling issues, rather than the specific challenge SESMG addresses.  Given your focus on "low threshold entry", "no programming skills" and "advanced modelling features", I think you should properly **motivate** this in the Statement of Need section.  At present, you just reiterate the functionality without providing a rationale or justification for them.

- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

  - [ ] You do not include any references to other software designed for urban energy system modelling.

- [ ] A list of key references, including other software addressing related needs. Note that the references should include full names of venues, e.g., journals and conferences, not abbreviations only understood in the context of a specific discipline.

  - [ ] Please add references to other software the provides similar functionality and explain what differentiates your software.

- [ ] Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.

  - [ ] You have included information on previous papers and projects that have used the software, but this is not mentioned in the paper. [optional]

- [x] Acknowledgement of any financial support.
