### 2023-07-04 bb96e40
Update paper.md
Update regarding the following reviewer suggestion: "I think you could more clearly outline what functionality the software provides e.g. a browser-based graphical user interface, spreadsheets to provide data entry, comprehensive documentation on how to use it"
