### 2023-07-04 42d1283
docu(paper.md): Adaptions regarding reviewer suggestions
Adaptions regarding the following reviewer suggestions:
- "You do not include any references to other software designed for urban energy system modelling."
-  "Please add references to other software the provides similar functionality and explain what differentiates your software."
