### 2023-07-04 09acb64
docu(paper.md): adaptions regarding reviewer suggestions
Adaptions regarding the following reviewer suggestion: "The current text in this part of the paper is more focussed on general energy modelling issues, rather than the specific challenge SESMG addresses. Given your focus on "low threshold entry", "no programming skills" and "advanced modelling features", I think you should properly motivate this in the Statement of Need section. At present, you just reiterate the functionality without providing a rationale or justification for them."
