### 2023-07-05 e97f373
docu(paper.md): Adaption regarding reviewer comments
Adaption regarding the following reviewer comment: "This includes many technical terms that are not defined e.g. "sector coupling"
