### 2023-07-05 fda0ac3
docu(paper.md): typo
