### 2023-07-05 2a7e79e
docu(paper.md): language corrections
