### 2023-07-06 b03ea2e
docu(paper.md): typo
