### 2023-07-06 b590237
docu(paper.md): typo
