### 2023-07-06 ebd2dea
docu(paper.md): typo
