+++ docs/publications/paper.md
@@ -54,7 +54,7 @@ The SESMG enables the optimization of multi-sectoral and spatial synergies of en
 
 The target group of the SESMG are planners of (urban) energy systems and researchers in the field of energy engineering. As it is required for the application of the SESMG and the interpretation of the results, users are required to have a certain basic knowledge of energy systems and energy engineering. Compared to other tools for the modeling and optimization of urban energy systems, as they have been listed by Klemm and Vennemann [@Klemm.Review], the SESMG provides several advantages regarding user-friendliness due to
 
-* beeing available under an **open-source license**, 
+* being available under an **open-source license**, 
 
 * applicability **without any programming knowledge**,
 
