### 2023-07-10 SESMG/SESMG#194 @chrklemm
[JOSS] Reviewer Comments
**Code of Conduct**
- [x] Hereby I consent to the [CODE_OF_CONDUCT](https://github.com/SESMG/SESMG/blob/master/CODE_OF_CONDUCT.md)

Following are reviewer comments by @nick-gorman from a [JOSS-Review](https://github.com/openjournals/joss-reviews/issues/5519):

> @chrklemm SESMG looks like great software. I really like the idea of making the omeof capabilities available to users who don't use python by creating a standard spreadsheet input format.
>
> Here is my review of the documentation and paper as per the checklist:
>
> Functionality:
>
> * I haven't been able to run the file model_definition_example.xlsx to completion. I'll raise an issue about this separately. Just as an aside, examples that run quickly are more useful for helping new users learn about the tool.
>
> Documentation:
>
> * I can't find a clear statement of need (please direct me in the right direction if one exists). Further to this, although not required by the JOSS review I think the read-the-docs pages lack a good introduction to orient the reader. At least a statement of need should be added, and the authors should consider adding an intro to the read-the-docs page.
> * Again, while technically the documentation includes an example as part of the application section this isn't broken out separately in a way that makes it easy to understand holistically, i.e. explain the energy system, explain the problem that is being solved in the example, walk the user through how this is done with the software. Not 100% necessary for the JOSS review but should be considered by the authors.
> * Also, the examples folder on GitHub appears empty? (https://github.com/SESMG/SESMG/tree/master/examples)
> * Functionality documentation, I don't think the descriptions of the various modelling capability are detailed/clear enough. This would probably be improved with an example that uses each type of functionality, i.e. an example that uses clustering that demonstrates how inputs affect the clustering (for example, it isn't clear to me what selecting hours for clustering will actually do).
>
> Software paper:
>
> * Not sure that the statement of need addresses who the target audience is.
> * Other commonly used packages are not addressed in the paper.
