### 2023-07-10 #5519 @chrklemm
Dear @nick-gorman, thank you for your helpful comments! I have moved the comments to a [sepparate issue](https://github.com/SESMG/SESMG/issues/194) where we will answer them. I hope this is ok for you.

Best regards,
@chrklemm
