### 2023-07-10 dfd11b9
docu(paper.md): language corrections with special focus on british/american english
