### 2023-07-12 7e3d312
docu(examples): removed obsolete folder
