### 2023-07-12 44313ef
docu(examples): added example section
