### 2023-07-12 df8a1ff
docu(readme.md): updated link to example repository
