### 2023-07-12 f7c0db2
docu(README.md): typo
