### 2023-07-13 b84cc52
docu(main page): restructured main page
