+++ docs/publications/paper.bib
@@ -156,11 +156,11 @@
 }
 
 @misc{oemof,
-  title = {Oemof Solph},
-  author = {{oemof-developer-group}},
-  year = {2023},
-  abstract = {A model generator for energy system modelling and optimisation (LP/MILP).},
-  howpublished = {https://github.com/oemof/oemof-solph}
+author = {Krien, Uwe and Kaldemeyer, Cord and Günther, Stephan and Schönfeldt, Patrik and Simon, Hilpert and Launer, Jann and Röder, Johannes and Möller, Caroline and Kochems, Johannes and Huyskens, Hendrik and @steffenGit and Schachler, Birgit and Pl, Franzi and Sayadi, Saeed and Duc, Pierre-François and Endres, Julian and Büllesbach, Fabian and Fuhrländer, David and @gplssm and Francesco, Witte and Kassing, Phillip and Zolotarevskaia, Ekatarina and Berendes, Sarah and Lancien, Bryan and Developer, Anonymous and Developer, Anonymous and Developer, Anonymous and Schürmann, Lennart and Developer, Anonymous and Delfs, Jens-Olaf and Developer, Anonymous and Developer, Anonymous and Smalla, Tjark and Developer, Anonymous and Wolf, Jakob and Developer, Anonymous and Gaudchau, Elisa and Developer, Anonymous and Rohrer, Tobi},
+doi = {10.5281/zenodo.596235},
+license = {MIT},
+title = {{oemof.solph}},
+url = {https://github.com/oemof/oemof-solph/}
 }
 
 @article{Klemm.Review,
