+++ docs/index.rst
@@ -1,18 +1,10 @@
 The Spreadsheet Energy System Model Generator 
 =================================================
 
-The **SESMG** is an energy system model generator with the focus on the optimization of urban energy systems, which can, however, also be used for the modeling of other types of energy systems. The **SESMG** is based on the 'Open Energy Modelling Framework' (oemof) and comes, compared to other modeling tools with advantages regarding user-friendliness, as
- 
- * the model definition is based on spreadsheets, therefore no programming skills are required for the entire modeling process,
- * urban energy system models with any size can automatically conceptualized,
- * visualization of complex results are automatically created in the form of system graphs, Pareto-fronts, energy amounts, capacity diagrams, and many more, as well as
- * a set of standard (but still customizable) parameters are given, including detailed descriptions and references.
- 
-Furthermore, the **SESMG** comes with important modeling methods, enabling holistic modeling of spatially high resolution modeling of mixed-use multi energy systems, such as
- 
- * considering the multi-energy system (MES) approach
- * applying multi-objective optimization by using the epsilon-constraint-method, as well as
- * providing several methods for model-based reduction of computational requirements (run-time and RAM).
+What is the SESMG?
+-------------------------
+
+The Spreadsheet Energy System Model Generator (SESMG) is a tool for modeling and optimizing energy systems with a focus on urban systems. The SESMG is easyily accessible as it comes with a browser-based graphical user interface, spreadsheets to provide data entry, and detailed documentation on how to use it. **Programming skills are not required** for the installation or application of the tool. The SESMG includes **advanced modeling features** such as the application of the **multi-energy system (MES)** approach, **multi-objective optimization**, model-based methods for **reducing computational requirements**, and **automated conceptualization and result processing** of urban energy systems with high spatial resolution. Due to its accessibility and the applied modeling methods, urban energy systems can be modeled and optimized with comparatively low effort. [#FN1]_
 
 
 .. figure:: images/SESMG_principle.png
@@ -20,11 +12,50 @@ Furthermore, the **SESMG** comes with important modeling methods, enabling holis
    :alt: SESMG_principle.png
    :align: center
 
-   From an input spreadsheet to interactive modelling results: The Spreadsheet Energy System Model Generator
+   From an input spreadsheet to interactive modeling results: The Spreadsheet Energy System Model Generator
 
+
+Why is the SESMG useful?
+-------------------------------
+
[…]
