### 2023-07-13 518a08a
docs(paper.md): added gui info
