### 2023-07-13 5a1a972
docu(paper.md): typo
