### 2023-07-14 fb4e517
docu(index.rst): minor adaption
