+++ .github/workflows/build.yml
@@ -42,8 +42,8 @@ jobs:
           ls -al
           export VER="0.5.0"
           echo "VER=$VER" >> $GITHUB_ENV
-        #  mkdir dist_actions
-        # ci/macos-deploy.sh
+          mkdir dist_actions
+          ci/macos-deploy.sh
         #  ls -al dist_actions
       - name: Deploy
         uses: actions/upload-artifact@v2
