+++ .github/workflows/build.yml
@@ -43,13 +43,15 @@ jobs:
           export VER="0.5.0"
           echo "VER=$VER" >> $GITHUB_ENV
           mkdir dist_actions
-          ci/macos-deploy.sh
+          pip3 install dmgbuild
+          python3 -m dmgbuild -s start_script.py "SESMG" dist_actions/SESMG_${VER}.dmg
+        #  ci/macos-deploy.sh
         #  ls -al dist_actions
       - name: Deploy
         uses: actions/upload-artifact@v2
         with:
           name: SESMG_${{ env.VER }}.dmg
           path: |
-            program_files/SESMG_${{ env.VER }}.dmg
+            dist_actions/SESMG_${{ env.VER }}.dmg
 
  
