+++ .github/workflows/build.yml
@@ -43,12 +43,9 @@ jobs:
           export VER="0.5.0"
           echo "VER=$VER" >> $GITHUB_ENV
           mkdir dist_actions
-          pip3 install dmgbuild
-          cd dist 
-          ls
-          python3 -m dmgbuild -s start_script.py "SESMG" dist_actions/SESMG_${VER}.dmg
-        #  ci/macos-deploy.sh
-        #  ls -al dist_actions
+          ci/macos-deploy.sh
+          ls -al dist_actions
+          
       - name: Deploy
         uses: actions/upload-artifact@v2
         with:
