+++ ci/appdmg.py
@@ -1,0 +1,166 @@
+# -*- coding: utf-8 -*-
+import os.path
+
+import plistlib
+
+application = defines.get('app', 'dist/SESMG.app')
+appname = os.path.basename(application)
+
+def read_plist(path):
+    with open(path, 'rb') as f:
+        return plistlib.load(f)
+
+def icon_from_app(app_path):
+    plist_path = os.path.join(app_path, 'Contents', 'Info.plist')
+    plist = read_plist(plist_path)
+    icon_name = plist['CFBundleIconFile']
+    icon_root, icon_ext = os.path.splitext(icon_name)
+    if not icon_ext:
+        icon_ext = '.icns'
+    icon_name = icon_root + icon_ext
+    return os.path.join(app_path, 'Contents', 'Resources', icon_name)
+
+# Volume format (see hdiutil create -help)
+format = defines.get('format', 'UDZO')
+
+# Compression level (if relevant)
+compression_level = 9
+
+# Volume size
+size = defines.get('size', None)
+
+# Files to include
+files = [
+    application,
+    #'syncplay/resources/.macOS_readme.pdf'
+]
+
+# Symlinks to create
+symlinks = {
+    'Applications': '/Applications',
+    #'Read Me': '.macOS_readme.pdf'
+}
+
+# Volume icon
+#
+# You can either define icon, in which case that icon file will be copied to the
+# image, *or* you can define badge_icon, in which case the icon file you specify
+# will be used to badge the system's Removable Disk icon
+#
+# icon = '/path/to/icon.icns'
+badge_icon = icon_from_app(application)
+
+# Where to put the icons
+icon_locations = {
+    appname: (90, 110),
+    'Applications': (410, 110),
+    #'Read Me': (250, 240),
+}
+
+# .. Window configuration ......................................................
+
+# Background
+#
+# This is a STRING containing any of the following:
+#
+#    #3344ff          - web-style RGB color
+#    #34f             - web-style RGB color, short form (#34f == #3344ff)
+#    rgb(1,0,0)       - RGB color, each value is between 0 and 1
+#    hsl(120,1,.5)    - HSL (hue saturation lightness) color
+#    hwb(300,0,0)     - HWB (hue whiteness blackness) color
+#    cmyk(0,1,0,0)    - CMYK color
+#    goldenrod        - X11/SVG named color
+#    builtin-arrow    - A simple built-in background with a blue arrow
+#    /foo/bar/baz.png - The path to an image file
+#
+# The hue component in hsl() and hwb() may include a unit; it defaults to
+# degrees ('deg'), but also supports radians ('rad') and gradians ('grad'
+# or 'gon').
+#
+# Other color components may be expressed either in the range 0 to 1, or
+# as percentages (e.g. 60% is equivalent to 0.6).
+#background = 'syncplay/resources/macOS_dmg_bkg.tiff'
+
+show_status_bar = False
+show_tab_view = False
+show_toolbar = False
+show_pathbar = False
+show_sidebar = False
+sidebar_width = 180
+
+# Window position in ((x, y), (w, h)) format
+window_rect = ((100, 100), (500, 400))
+
+# Select the default view; must be one of
+#
+#    'icon-view'
+#    'list-view'
+#    'column-view'
+#    'coverflow'
+#
+default_view = 'icon-view'
+
+# General view configuration
+show_icon_preview = False
+
+# Set these to True to force inclusion of icon/list view settings (otherwise
+# we only include settings for the default view)
+include_icon_view_settings = 'auto'
+include_list_view_settings = 'auto'
+
+# .. Icon view configuration ...................................................
+
+arrange_by = None
+grid_offset = (0, 0)
[…]
