+++ ci/macos-deploy.sh
@@ -10,4 +10,4 @@ python3 ci/macos_app_adhoc_sign.py
 pip3 install dmgbuild
 #mv SESMG/resources/macOS_readme.pdf SESMG/resources/.macOS_readme.pdf
 
-python3 -m dmgbuild -s appdmg.py "SESMG" dist_actions/SESMG_${VER}.dmg
+python3 -m dmgbuild -s ci/appdmg.py "SESMG" dist_actions/SESMG_${VER}.dmg
