+++ ci/macos-deploy.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-set -ex
+#set -ex
 
 #python3 ci/macos_app_cleaner.py
 python3 ci/macos_app_arch_check.py
