### 2023-07-14 84c8150
Update macos_app_arch_check.py
