### 2023-07-14 7d4baee
Update macos_app_adhoc_sign.py
