+++ .github/workflows/build.yml
@@ -42,6 +42,9 @@ jobs:
           ls -al
           export VER="0.5.0"
           echo "VER=$VER" >> $GITHUB_ENV
+          cd dist
+          ls
+          cd ..
           mkdir dist_actions
           . ci/macos-deploy.sh
           ls -al dist_actions
