+++ setup.py
@@ -12,7 +12,8 @@ with open("requirements.txt") as f:
 
 OPTIONS = {
     'argv_emulation': True,
-    'includes': ["program_files", "streamlit"]
+    'includes': ["program_files"],
+    "packages": ["streamlit"]
 }
 
 setuptools.setup(
