+++ setup.py
@@ -27,7 +27,7 @@ setuptools.setup(
     #requires=install_requires,
     app=["start_script.py"],
     options={'py2app': OPTIONS},
-    setup_requires=["py2app", "streamlit"],
+    setup_requires=install_requires,
     install_requires=install_requires,
     classifiers=[
         # complete classifier list:
