+++ .github/workflows/build.yml
@@ -22,6 +22,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           pip3 install -U setuptools wheel pip
+          pip3 install modulegraph macholib altgraph
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt
       - name: Install py2app
