+++ .github/workflows/build.yml
@@ -21,23 +21,12 @@ jobs:
           pip3 --version
       - name: Install Python dependencies
         run: |
-          pip3 install -U setuptools wheel pip
-          pip3 install modulegraph macholib altgraph
+          pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt
-      - name: Install py2app
-        run: |
-          pip3 install py2app
-          python3 setup.py build
-          python3 setup.py install
-      - name: Check Python dependencies
-        run: |
-          echo $DYLD_LIBRARY_PATH
-          echo $DYLD_FRAMEWORK_PATH
-          # python3 -c 'from distutils.sysconfig import get_config_var; print(get_config_var("LDLIBRARY"))'
       - name: Build
         run: |
-          python3 setup.py py2app
+          pyinstaller --onefile start_script.spec
       - name: Prepare for deployment
         run: |
           ls -al
