+++ .github/workflows/build.yml
@@ -28,6 +28,7 @@ jobs:
           pip3 install -r requirements.txt
       - name: Build
         run: |
+          cd lib 
           ls
           pyinstaller start_script.spec
       - name: Prepare for deployment
