+++ .github/workflows/build.yml
@@ -28,8 +28,10 @@ jobs:
           pip3 install -r requirements.txt
       - name: Build
         run: |
-          cd lib 
+          cd lib/python3.9 
           ls
+          cd ..
+          cd ..
           pyinstaller start_script.spec
       - name: Prepare for deployment
         run: |
