+++ .github/workflows/build.yml
@@ -28,8 +28,9 @@ jobs:
           pip3 install -r requirements.txt
       - name: Build
         run: |
-          cd lib/python3.9 
+          cd lib/python3.9/site-packages 
           ls
+          cd ..
           cd ..
           cd ..
           pyinstaller start_script.spec
