+++ .github/workflows/build.yml
@@ -21,11 +21,9 @@ jobs:
           pip3 --version
           python3 -m venv .
           . bin/activate
-      - name: Install Python dependencies
-        run: |
           pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
-          pip3 install -r requirements.txt
+          pip3 install -r requirements.txt  
       - name: Build
         run: |
           cd lib/python3.9/site-packages 
