+++ .github/workflows/build.yml
@@ -24,8 +24,6 @@ jobs:
           pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt  
-      - name: Build
-        run: |
           pyinstaller start_script.spec
       - name: Prepare for deployment
         run: |
