+++ .github/workflows/build.yml
@@ -24,7 +24,7 @@ jobs:
           pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt  
-          pyinstaller start_script.spec
+          pyinstaller SESMG.spec
       - name: Prepare for deployment
         run: |
           ls -al
