### 2023-07-18 e19053a
Update 02.02.06_examples.rst
Adding example description
