+++ docs/02.02.06_examples.rst
@@ -1,13 +1,181 @@
 Examples
 ==============
-For the SESMG, a number of examples can be found in this `this directory <https://github.com/SESMG/SESMG-Examples>`_. Please make sure that the version indexed with the file/folder name matches your version of the SESMG. In the following, some examples are explained in more detail.
+For the SESMG, a number of examples can be found in `this directory <https://github.com/SESMG/SESMG-Examples>`_. Please make sure that the version indexed with the file/folder name matches your version of the SESMG. In the following, some examples are explained in more detail.
 
-Example 1
+Example 1: Simulation and optimization of a single family house
 ---------
+The template for this exercise is this Excel sheet. You can find the completed Excel sheet with the solutions here.
 
-Example 2
----------
+Example 1a: Simulation of a single family house
+~~~~~~~~~~
+The Müller family has three children and lives in a single-family house with 180 m\ :sup:`2`. They are planning to adapt the energy supply of their house. The entire heating demand of the family is currently provided by a gas heating system. To cover their electricity needs, the Müllers decided last year to install a PV system on the roof of the house. The rest of their electricity needs are met conventionally by connecting to the grid.
 
-Example 3
+Known data:
+^^^^^^^^^^^^^^^^^^^^^^^^^^
++---------------------+---------------+
+| Data                | Value         |
++=====================+===============+
+| Electricity Demand  | 5,100 kWh     |
++---------------------+---------------+
+| Heat Demand         | 32,000 kWh    |
++---------------------+---------------+
+| PV System           | 5 kWp         |
++---------------------+---------------+
+Simplification:
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+- The PV system should be regarded as a new investment, as only a small portion of the depreciated to a small extent
+- Using the time series simplification Slicing A with 4 days (for hole example 1)
+.. image:: C:/Users/BenjaminBlankenstein/OneDrive/Dokumente/Praxisphase/Screenshots/time_series_simplifications.png
+   :width: 300px
+   :align: left
+
+Simulation task: 
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Simulate the given energy system
+- Activate the given components via the column "active" as binary input. 0 = off, 1 = on
+
+.. attention:: 
+	- After each change you have to save and upload the model definition again
+
+Example 1b: Cost optimization of the energy system
+~~~~~~~~~~
+From the results of the simulation of the single-family house, it can be concluded that it makes sense to incorporate additional technologies in order to reduce the cost of energy supply. In this example, the following technologies are available:
+
+- Air source heat pump (ASHP) with max. 20 kW 
+- Ground source heat pump (GCHP) with max. 15 kW
+- Battery storage with max. 10 kWh
+- Thermal storage with max. 20 kWh
+
+Optimization task: 
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Adjusts the templates of the individual technologies in the model definition accordingly and supplements them in the system
+
+.. attention:: 
+	- It is possible to purchase electricity for heat pumps at a discounted rate
+	- Simulation (Example 1a): Min. and max. investment capacity identical
+	- Optimization (Example 1b): Interval between min. and max. investment capacity
+
+Example 1c: Pareto optimization of a single family house
+~~~~~~~~~~
+The Müller family has heard that the emissions caused by energy systems can be significantly reduced by low additional costs.
+
+Pareto-optimization task: 
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Execute a Pareto optimization of the energy system
+- Calculate the cost and emission minimums, as well three other Pareto points
+- Select the points in such a way that they are as meaningful as possible.
+
+.. attention:: 
+	- 0 or 0 % represents the cost minimum, since 0 % of the possible emission reduction is exhausted
[…]
