### 2023-07-18 0ecc990
Update 02.02.06_examples.rst
Updatet .rst with weblinks and screenshot
