### 2023-07-18 45e845b
docu(02.02.06_examples.rst): added links to examples
Added links and screenshot
