+++ docs/02.02.06_examples.rst
@@ -4,7 +4,10 @@ For the SESMG, a number of examples can be found in `this directory <https://git
 
 Example 1: Simulation and optimization of a single family house
 ---------
-The template of the excel sheet for this exercise you can find `here <https://github.com/SESMG/SESMG-Examples/blob/main/SESMG-v0.5.x-Examples/Task_1_template_model_definition.xlsx>`_. You can find the completed excel sheet with solutions `here <https://github.com/SESMG/SESMG-Examples/blob/main/SESMG-v0.5.x-Examples/Task_1_solution_model_definition.xlsx>`_.
+The template of the model definition (.xlsx) for this exercise you can find `here <https://github.com/SESMG/SESMG-Examples/blob/main/SESMG-v0.5.x-Examples/Task_1_template_model_definition.xlsx>`_. You can find the completed
+model definition with solutions `here <https://github.com/SESMG/SESMG-Examples/blob/main/SESMG-v0.5.x-Examples/Task_1_solution_model_definition.xlsx>`_. This example is intended as an introduction and has an exercise
+character.
+
 
 Example 1a: Simulation of a single family house
 ~~~~~~~~~~
@@ -15,9 +18,9 @@ Known data:
 +---------------------+---------------+
 | Data                | Value         |
 +=====================+===============+
-| Electricity Demand  | 5,100 kWh     |
+| Electricity Demand  | 5 100 kWh     |
 +---------------------+---------------+
-| Heat Demand         | 32,000 kWh    |
+| Heat Demand         | 32 000 kWh    |
 +---------------------+---------------+
 | PV System           | 5 kWp         |
 +---------------------+---------------+
@@ -25,7 +28,7 @@ Known data:
 Simplification:
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 - The PV system should be regarded as a new investment, as only a small portion of the depreciated to a small extent
-- Using the time series simplification Slicing A with 4 days (for hole example 1)
+- Using the time series simplification "Slicing A" with 4 days (for hole example 1)
    
 .. figure:: ../docs/images/manual/Examples/time_series_simplification.png
    :width: 20 %
@@ -35,10 +38,12 @@ Simplification:
 Simulation task: 
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 - Simulate the given energy system
+	- To find out how to start a simulation, click `here <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.02.00_application.html#main-application>`
 - Activate the given components via the column "active" as binary input. 0 = off, 1 = on
 
 .. attention:: 
-	- After each change you have to save and upload the model definition again
+- After each change you have to save and upload the model definition again
+- Simulation: Min. and max. investment capacity identical. You use the same button for simulation as for optimization.
 
 Example 1b: Cost optimization of the energy system
 ~~~~~~~~~~
@@ -52,11 +57,12 @@ From the results of the simulation of the single-family house, it can be conclud
 Optimization task: 
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 - Adjusts the templates of the individual technologies in the model definition accordingly and supplements them in the system
+	- To find out how to start a optimazation, click `here <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.02.00_application.html#main-application>`
 
 .. attention:: 
-	- It is possible to purchase electricity for heat pumps at a discounted rate
-	- Simulation (Example 1a): Min. and max. investment capacity identical
-	- Optimization (Example 1b): Interval between min. and max. investment capacity
[…]
