### 2023-07-18 849c84f
docu(02.02.06_examples.rst): correcting text
correcting text
