### 2023-07-18 e59fecf
docu(02.02.06_examples.rst): correcting note boxes
correcting note boxes
