+++ docs/02.02.06_examples.rst
@@ -3,18 +3,18 @@ Examples
 For the SESMG, a number of examples can be found in `this directory <https://github.com/SESMG/SESMG-Examples>`_. Please make sure that the version indexed with the file/folder name matches your version of the SESMG. In the following, some examples are explained in more detail.
 
 Example 1: Simulation and optimization of a single family house
----------
+---------------------------------------------------------------
 The template of the model definition (.xlsx) for this exercise you can find `here <https://github.com/SESMG/SESMG-Examples/blob/main/SESMG-v0.5.x-Examples/Task_1_template_model_definition.xlsx>`_. You can find the completed
 model definition with solutions `here <https://github.com/SESMG/SESMG-Examples/blob/main/SESMG-v0.5.x-Examples/Task_1_solution_model_definition.xlsx>`_. This example is intended as an introduction and has an exercise
 character.
 
 
 Example 1a: Simulation of a single family house
-~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 The Müller family has three children and lives in a single-family house with 180 m\ :sup:`2`. They are planning to adapt the energy supply of their house. The entire heating demand of the family is currently provided by a gas heating system. To cover their electricity needs, the Müllers decided last year to install a PV system on the roof of the house. The rest of their electricity needs are met conventionally by connecting to the grid.
 
 Known data:
-^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^
 +---------------------+---------------+
 | Data                | Value         |
 +=====================+===============+
@@ -26,7 +26,7 @@ Known data:
 +---------------------+---------------+
 
 Simplification:
-^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^
 - The PV system should be regarded as a new investment, as only a small portion of the depreciated to a small extent
 - Using the time series simplification "Slicing A" with 4 days (for hole example 1)
    
@@ -36,17 +36,18 @@ Simplification:
    :align: center   
 
 Simulation task: 
-^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^^^^^^^^
 - Simulate the given energy system
-	- To find out how to start a simulation, click `here <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.02.00_application.html#main-application>`
+	- To find out how to start a simulation, click `here <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.02.00_application.html#main-application>`_
 - Activate the given components via the column "active" as binary input. 0 = off, 1 = on
 
-.. attention:: 
-- After each change you have to save and upload the model definition again
-- Simulation: Min. and max. investment capacity identical. You use the same button for simulation as for optimization.
+.. attention::
+
+	- After each change you have to save and upload the model definition again
+	- Simulation: Min. and max. investment capacity identical. You use the same button for simulation as for optimization.
 
 Example 1b: Cost optimization of the energy system
-~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 From the results of the simulation of the single-family house, it can be concluded that it makes sense to incorporate additional technologies in order to reduce the cost of energy supply. In this example, the following technologies are available:
 
 - Air source heat pump (ASHP) with max. 20 kW 
[…]
