### 2023-07-18 SESMG/SESMG#194 @chrklemm
Dear @nick-gorman, thank you very much for your positive feedback! We have incorporated your suggestions within the ‘issue-194’-branch and will push it to the main branch as soon as you agree to the changes.

> Here is my review of the documentation and paper as per the checklist:

>

> Functionality:

>

> * I haven't been able to run the file model_definition_example.xlsx to completion. I'll raise an issue about this separately. Just as an aside, examples that run quickly are more useful for helping new users learn about the tool.

>

We have added two simple step-by-step tutorial examples that introduce new users to the various functionalities of the SESMG (mainly incorporated with commit e19053a).

> Documentation:

>

> * I can't find a clear statement of need (please direct me in the right direction if one exists). Further to this, although not required by the JOSS review I think the read-the-docs pages lack a good introduction to orient the reader. At least a statement of need should be added, and the authors should consider adding an intro to the read-the-docs page.

We totally restructured the starting page of the documentation. It now consists of the three sections of “What is the SESMG?”, ”Why is the SESMG useful?”, ”How is the documentation structured?” (see commit [b84cc52](https://github.com/SESMG/SESMG/commit/b84cc52c014d885575ad9c7ea39f012b948489c6))

> * Again, while technically the documentation includes an example as part of the application section this isn't broken out separately in a way that makes it easy to understand holistically, i.e. explain the energy system, explain the problem that is being solved in the example, walk the user through how this is done with the software. Not 100% necessary for the JOSS review but should be considered by the authors.

We have added two simple step-by-step tutorial examples that introduce new users to the various functionalities of the SESMG (see commit e19053a).

> * Also, the examples folder on GitHub appears empty? (https://github.com/SESMG/SESMG/tree/master/examples)

We have moved the examples section into an own sesmg-repository for the sake of tidyness. However, we made the following changes to make this clearer:

- We have removed the empty folder in the main repository (commit [7e3d312](https://github.com/SESMG/SESMG/commit/7e3d312fc8a4fcba6af0de73011b02570af5bb2f)).

- We have created an example chapter in the documentation that refers to the example repository (commit [44313ef).](https://github.com/SESMG/SESMG/commit/44313ef75bfc31eec311308be1d1c3df06de623e)

- We have revised the structure of the example-repository (commit [9523a2d).](https://github.com/SESMG/SESMG-Examples/commit/9523a2dbf50778616ff7045a1378a4bf2d46f42d)

- We have added two simple step-by-step tutorial examples that introduce new users to the various functionalities of the SESMG (mainly incorporated with commit e19053a).

> * Functionality documentation, I don't think the descriptions of the various modelling capability are detailed/clear enough. This would probably be improved with an example that uses each type of functionality, i.e. an example that uses clustering that demonstrates how inputs affect the clustering (for example, it isn't clear to me what selecting hours for clustering will actually do).

We have added two simple step-by-step tutorial examples that introduce new users to the various functionalities of the SESMG (mainly incorporated with commit e19053a).

>

> Software paper:

>

> * Not sure that the statement of need addresses who the target audience is.

You are totally right, that point was missing. Based on a reviewer comment of the other reviewer, we have already incorporated that point with commit e97f373.

> * Other commonly used packages are not addressed in the paper.

Based on a comment of the other reviewer, we have incorporated this point: We referenced a complete list of other software and have made the advantages of the SESMG more clear (see commit 42d1283).
