+++ .github/workflows/build.yml
@@ -70,7 +70,7 @@ jobs:
           pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt  
-          pyinstaller SESMG_windows.spec
+          python3 - m pyinstaller SESMG_windows.spec
           cd dist
           ls
  
