+++ .github/workflows/build.yml
@@ -45,33 +45,5 @@ jobs:
           path: |
             dist/SESMG.dmg
             
-  windows:
-    name: Build on windows
-    runs-on: windows-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-
-      - name: Setup Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: '3.9'
-          architecture: x86
-          
-      - name: Check Python install
-        run: |
-          which python3
-          python3 --version
-          which pip3
-          pip3 --version
-          python3 -m venv .
-          ls
-          Scripts/activate.bat
-          pip3 install -U setuptools wheel pip pyinstaller
-          pip3 install twisted[tls] appnope requests certifi
-          pip3 install -r requirements.txt  
-          python3 - m pyinstaller SESMG_windows.spec
-          ls
-          cd dist
-          ls
+  
  
