### 2023-07-19 9886740
Revert "Update build.yml"
This reverts commit 6153ccbada84bbb9015531fc21068d53822e5549.
