+++ .github/workflows/build.yml
@@ -64,7 +64,7 @@ jobs:
           pip3 --version
           python3 -m venv .
           ls
-          Scripts/activate.bat
+          start /b Scripts\activate.bat
           pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt  
