### 2023-07-19 88cca2a
Revert "Update build_windows.yml"
This reverts commit 2345ac40846d9a1b6ac408a20f17414827ccb5a0.
