+++ .github/workflows/build_windows.yml
@@ -1,0 +1,36 @@
+name: Build
+on: [push, pull_request]
+
+jobs:
+  windows:
+      name: Build on windows
+      runs-on: windows-latest
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v2
+
+        - name: Setup Python
+          uses: actions/setup-python@v2
+          with:
+            python-version: '3.9'
+            architecture: x86
+
+        - name: Check Python install
+          run: |
+            which python3
+            python3 --version
+            which pip3
+            pip3 --version
+            python3 -m venv .
+            ls
+            Scripts/activate.bat
+            Scripts/pip3 install -U setuptools wheel pip pyinstaller
+            Scripts/pip3 install twisted[tls] appnope requests certifi
+            Scripts/pip3 install -r requirements.txt  
+            Scripts/python3 -m pyinstaller SESMG_windows.spec
+            ls
+            cd dist
+            ls
+
+
+ 
