+++ .github/workflows/build_windows.yml
@@ -22,8 +22,10 @@ jobs:
             which pip3
             pip3 --version
             python3 -m venv .
-            ls
             Scripts/activate.bat
+            cd Scripts
+            ls
+            cd ..
             Scripts/pip3 install -U setuptools wheel pip pyinstaller
             Scripts/pip3 install twisted[tls] appnope requests certifi
             Scripts/pip3 install -r requirements.txt  
