+++ .github/workflows/build_windows.yml
@@ -41,7 +41,7 @@ jobs:
             cd Scripts
             ls
             cd ..
-            Scripts/pip3 install -U setuptools wheel pyinstaller
+            Scripts/pip3 install -U setuptools wheel pyinstaller subprocess
             Scripts/pip3 install twisted[tls] appnope requests certifi
             Scripts/pip3 install -r requirements.txt  
             Scripts/python3 -m pyinstaller SESMG_windows.spec
