+++ .github/workflows/build_windows.yml
@@ -41,6 +41,7 @@ jobs:
             cd Scripts
             ls
             cd ..
+            Scripts/pip3 install tables --no-binary all
             Scripts/pip3 install -U setuptools wheel pyinstaller subprocess
             Scripts/pip3 install twisted[tls] appnope requests certifi
             Scripts/pip3 install -r requirements.txt  
