+++ .github/workflows/build_windows.yml
@@ -41,8 +41,9 @@ jobs:
             cd Scripts
             ls
             cd ..
+            Scripts/pip3 install --upgrade setuptools
             Scripts/pip3 install tables<3.0 --no-binary all
-            Scripts/pip3 install -U setuptools wheel pyinstaller subprocess
+            Scripts/pip3 install -U wheel pyinstaller subprocess
             Scripts/pip3 install twisted[tls] appnope requests certifi
             Scripts/pip3 install -r requirements.txt  
             Scripts/python3 -m pyinstaller SESMG_windows.spec
