+++ .github/workflows/build_windows.yml
@@ -35,6 +35,7 @@ jobs:
 
         - name: Check Python install
           run: |
+            $PYTHONPATH=/c/hostedtoolcache/windows/Python/3.9.13/x86/python3
             which python3
             python3 --version
             which pip3
