+++ .github/workflows/build_windows.yml
@@ -10,10 +10,9 @@ jobs:
           uses: actions/checkout@v2
 
         - name: Setup Python
-          uses: actions/setup-python@v2
+          uses: actions/setup-python@v4
           with:
             python-version: '3.9.13'
-            architecture: x86
             
         - name: Add msbuild to PATH
           uses: microsoft/setup-msbuild@v1.1
@@ -35,7 +34,6 @@ jobs:
 
         - name: Check Python install
           run: |
-            $PYTHONPATH=/c/hostedtoolcache/windows/Python/3.9.13/x86/python3
             which python3
             python3 --version
             which pip3
