+++ .github/workflows/build_windows.yml
@@ -47,7 +47,7 @@ jobs:
             Scripts/pip3 install pyinstaller
             Scripts/pip3 install twisted[tls] appnope requests certifi
             Scripts/pip3 install -r requirements.txt  
-            Scripts/python3 -m pyinstaller SESMG_windows.spec
+            pyinstaller SESMG_windows.spec
             ls
             cd dist
             ls
