+++ .github/workflows/build_windows.yml
@@ -40,14 +40,14 @@ jobs:
             pip3 --version
             python3 -m venv .
             Scripts/activate.bat
-            cd Scripts
-            ls
-            cd ..
             Scripts/pip3 install --upgrade setuptools
             Scripts/pip3 install pyinstaller
             Scripts/pip3 install twisted[tls] appnope requests certifi
             Scripts/pip3 install -r requirements.txt  
-            pyinstaller SESMG_windows.spec
+            cd Scripts
+            ls
+            cd ..
+            Scripts/pyinstaller SESMG_windows.spec
             ls
             cd dist
             ls
