+++ .github/workflows/build_macos.yml
@@ -24,7 +24,7 @@ jobs:
           pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt  
-          pyinstaller SESMG.spec
+          pyinstaller SESMG_macos.spec
       - name: Prepare for deployment
         run: |
           brew install create-dmg
