+++ .github/workflows/build_macos.yml
@@ -15,10 +15,7 @@ jobs:
           sudo installer -pkg pythonInstaller.pkg -target /
       - name: Check Python install
         run: |
-          which python3
-          python3 --version
-          which pip3
-          pip3 --version
+          brew install graphviz
           python3 -m venv .
           . bin/activate
           pip3 install -U setuptools wheel pip pyinstaller
