### 2023-07-27 76c3910
code(main_application): test to raise additional error information
