+++ .github/workflows/build_linux.yml
@@ -1,0 +1,37 @@
+name: Build
+on: [push, pull_request]
+
+jobs:
+  windows:
+      name: Build on windows
+      runs-on: linux-latest
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v2
+
+        - name: Setup Python
+          uses: actions/setup-python@v4
+          with:
+            python-version: '3.9.13'
+
+        - name: Check Python install
+          run: |
+            pip3 install virtualenv
+            python3 -m virtualenv .
+            . bin/activate
+            pip3 install --upgrade setuptools
+            pip3 install pyinstaller
+            pip3 install twisted[tls] appnope requests certifi
+            pip3 install -r requirements.txt  
+            pyinstaller SESMG_linux.spec
+            
+        - name: Deploy
+          uses: actions/upload-artifact@v2
+          with:
+            name: SESMG.sh
+            path: |
+              dist/SESMG.sh
+
+
+
+ 
