+++ .github/workflows/build_linux.yml
@@ -3,8 +3,8 @@ on: [push, pull_request]
 
 jobs:
   windows:
-      name: Build on windows
-      runs-on: linux-latest
+      name: Build on linux
+      runs-on: ubuntu-latest
       steps:
         - name: Checkout
           uses: actions/checkout@v2
