### 2023-07-31 #5519 @timtroendle
@openjournals/joss-eics, I will be out of office 7--25 August.
