### 2023-07-31 #5519 @chrklemm
Hello @timtroendle,

thank you for getting back in touch with us. We have processed most of the reviewer comments. They are waiting for confirmation from the reviewers that the issues can be closed:

https://github.com/SESMG/SESMG/issues/190 (responsible reviewer: @nick-gorman)
https://github.com/SESMG/SESMG/issues/193 (responsible reviewer: @willu47)
https://github.com/SESMG/SESMG/issues/194 (responsible reviewer: @nick-gorman)

The last issue is quite challenging for us, but we expect to complete it by the end of the week:

https://github.com/SESMG/SESMG/issues/188 (responsible reviewer: @willu47)

Best regards,
@chrklemm
