+++ .github/workflows/build_linux.yml
@@ -24,6 +24,18 @@ jobs:
             pip3 install twisted[tls] appnope requests certifi
             pip3 install -r requirements.txt  
             pyinstaller SESMG_linux.spec
+            sudo apt install ruby
+            gem install fpm --user-install
+            mkdir -p package/opt
+            mkdir -p package/usr/share/applications
+            mkdir -p package/usr/share/icons/hicolor/scalable/apps
+            cp -r dist/SESMG package/opt/SESMG
+            cp SESMG.desktop package/usr/share/applications
+            find package/opt/SESMG -type f -exec chmod 644 -- {} +
+            find package/opt/SESMG -type d -exec chmod 755 -- {} +
+            find package/usr/share -type f -exec chmod 644 -- {} +
+            chmod +x package/opt/SESMG/SESMG
+            fpm -C package -s dir -t deb -n "SESMG" -v 0.1.0 -p dist/SESMG.deb
             cd dist 
             ls 
             cd ..
@@ -31,9 +43,9 @@ jobs:
         - name: Deploy
           uses: actions/upload-artifact@v2
           with:
-            name: SESMG.sh
+            name: SESMG.deb
             path: |
-              dist/SESMG
+              dist/SESMG.deb
 
 
 
