+++ .github/workflows/build_linux.yml
@@ -19,6 +19,8 @@ jobs:
             pip3 install virtualenv
             python3 -m virtualenv .
             . bin/activate
+            sudo apt-get update 
+            sudo apt-get install libc6
             pip3 install --upgrade setuptools
             pip3 install pyinstaller
             pip3 install twisted[tls] appnope requests certifi
