+++ .github/workflows/build_linux.yml
@@ -4,7 +4,7 @@ on: [push, pull_request]
 jobs:
   windows:
       name: Build on linux
-      runs-on: ubuntu-20.4
+      runs-on: ubuntu-20.04
       steps:
         - name: Checkout
           uses: actions/checkout@v2
