### 2023-07-31 f055863
Delete ci directory
