+++ ci/macos_app_adhoc_sign.py
@@ -1,5 +1,0 @@
-from py2app.util import codesign_adhoc
-
-APPDIR = 'dist/SESMG.app'
-
-codesign_adhoc(APPDIR)
