+++ .github/workflows/build_macos.yml
@@ -21,7 +21,7 @@ jobs:
           pip3 install -U setuptools wheel pip pyinstaller
           pip3 install twisted[tls] appnope requests certifi
           pip3 install -r requirements.txt  
-          pyinstaller SESMG_macos.spec
+          pyinstaller .github/deploy/SESMG_macos.spec
       - name: Prepare for deployment
         run: |
           brew install create-dmg
