+++ .github/workflows/build.yml
@@ -114,7 +114,6 @@ jobs:
             path: |
               dist/SESMG.exe
 
-
             
   deploy:
     name: Push builds
