+++ .github/workflows/build_linux.yml
@@ -1,53 +1,0 @@
-name: Build
-on: [push, pull_request]
-
-jobs:
-  windows:
-      name: Build on linux
-      runs-on: ubuntu-20.04
-      steps:
-        - name: Checkout
-          uses: actions/checkout@v2
-
-        - name: Setup Python
-          uses: actions/setup-python@v4
-          with:
-            python-version: '3.9.13'
-
-        - name: Check Python install
-          run: |
-            pip3 install virtualenv
-            python3 -m virtualenv .
-            . bin/activate
-            sudo apt-get update 
-            sudo apt-get install libc6
-            pip3 install --upgrade setuptools
-            pip3 install pyinstaller
-            pip3 install twisted[tls] appnope requests certifi
-            pip3 install -r requirements.txt  
-            pyinstaller .github/deploy/SESMG_linux.spec
-            sudo apt-get install ruby-dev build-essential && sudo gem i fpm -f
-            mkdir -p package/opt
-            mkdir -p package/usr/share/applications
-            mkdir -p package/usr/share/icons/hicolor/scalable/apps
-            cp -r dist/SESMG package/opt/SESMG
-            cp .github/deploy/SESMG.desktop package/usr/share/applications
-            find package/opt/SESMG -type f -exec chmod 644 -- {} +
-            find package/opt/SESMG -type d -exec chmod 755 -- {} +
-            find package/usr/share -type f -exec chmod 644 -- {} +
-            chmod +x package/opt/SESMG/SESMG
-            fpm -C package -s dir -t deb -n "SESMG" -v 0.1.0 -p dist/SESMG.deb
-            cd dist 
-            ls 
-            cd ..
-            
-        - name: Deploy
-          uses: actions/upload-artifact@v2
-          with:
-            name: SESMG.deb
-            path: |
-              dist/SESMG.deb
-
-
-
- 
