+++ .github/workflows/build_macos.yml
@@ -1,50 +1,0 @@
-name: Build
-on: [push, pull_request]
-
-jobs:
-  macos:
-    name: Build for macOS
-    runs-on: macos-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-
-      - name: Setup Python
-        run: |
-          curl https://www.python.org/ftp/python/3.9.13/python-3.9.13-macosx10.9.pkg --output pythonInstaller.pkg
-          sudo installer -pkg pythonInstaller.pkg -target /
-      - name: Check Python install
-        run: |
-          python3 -m venv .
-          . bin/activate
-          pip3 install -U setuptools wheel pip pyinstaller
-          pip3 install twisted[tls] appnope requests certifi
-          pip3 install -r requirements.txt  
-          pyinstaller .github/deploy/SESMG_macos.spec
-      - name: Prepare for deployment
-        run: |
-          brew install create-dmg
-          mkdir -p dist/dmg
-          cp -r "dist/SESMG.app" dist/dmg
-          create-dmg \
-          --volname "SESMG" \
-          --hide-extension "SESMG.app" \
-          "dist/SESMG.dmg" \
-          "dist/dmg/"
-          
-      - name: Deploy
-        uses: actions/upload-artifact@v2
-        with:
-          name: SESMG.dmg
-          path: |
-            dist/SESMG.dmg
-            
-      - name: Push Project B
-        run: |
-           cd dist
-           git add SESMG.dmg
-           git config user.name deploy-action
-           git commit -am "Deployment of MacOS executable."
-           git push
-
- 
