+++ .github/workflows/semantic_versioning.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - master
+      - test
       
 permissions: 
   contents: read
