+++ .github/workflows/build.yml
@@ -115,7 +115,6 @@ jobs:
             name: SESMG.exe
             path: |
               dist/SESMG.exe
-
             
   deploy:
     name: Push builds
