+++ .github/workflows/semantic_versioning.yml
@@ -28,7 +28,7 @@ jobs:
   macos:
     name: Build for macOS
     runs-on: macos-latest
-    Needs: release
+    needs: release
     steps:
       - name: Checkout
         uses: actions/checkout@v2
