+++ .github/workflows/semantic_versioning.yml
@@ -19,6 +19,11 @@ jobs:
       - uses: actions/checkout@v2
         with:
           fetch-depth: 0
+          
+      - name: get current version 
+        run: |
+          current_release=$(python setup.py --version)
+          echo "current_release=v$current_release" >> $GITHUB_ENV
 
       - name: Python Semantic Release
         uses: relekang/python-semantic-release@master
@@ -32,12 +37,18 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v2
-
+      - name: get current version 
+        run : |
+          version=$(python setup.py --version)
+          echo "version=v$version" >> $GITHUB_ENV
+          
       - name: Setup Python
+        if: ${{ env.current_release != env.version }}
         run: |
           curl https://www.python.org/ftp/python/3.9.13/python-3.9.13-macosx10.9.pkg --output pythonInstaller.pkg
           sudo installer -pkg pythonInstaller.pkg -target /
       - name: Create Executable
+        if: ${{ env.current_release != env.version }}
         run: |
           python3 -m venv .
           . bin/activate
@@ -46,6 +57,7 @@ jobs:
           pip3 install -r requirements.txt  
           pyinstaller .github/deploy/SESMG_macos.spec
       - name: Prepare for deployment
+        if: ${{ env.current_release != env.version }}
         run: |
           brew install create-dmg
           mkdir -p dist/dmg
@@ -56,6 +68,7 @@ jobs:
           "dist/SESMG.dmg" \
           "dist/dmg/"
       - name: Upload Artifact
+        if: ${{ env.current_release != env.version }}
         uses: actions/upload-artifact@v3
         with:
           name: SESMG.dmg
@@ -67,14 +80,17 @@ jobs:
     needs: [release, macos]
     steps:
       - name: Checkout
+        if: ${{ env.current_release != env.version }}
         uses: actions/checkout@v2
 
       - name: Setup Python
+        if: ${{ env.current_release != env.version }}
         uses: actions/setup-python@v4
         with:
           python-version: '3.9.13'
 
       - name: Create Executable
+        if: ${{ env.current_release != env.version }}
         run: |
           pip3 install virtualenv
           python3 -m virtualenv .
@@ -87,6 +103,7 @@ jobs:
           pip3 install -r requirements.txt  
           pyinstaller .github/deploy/SESMG_linux.spec
       - name: Prepare for deployment
+        if: ${{ env.current_release != env.version }}
         run: |
           sudo apt-get install ruby-dev build-essential && sudo gem i fpm -f
           mkdir -p package/opt
@@ -100,6 +117,7 @@ jobs:
           chmod +x package/opt/SESMG/SESMG
           fpm -C package -s dir -t deb -n "SESMG" -v 0.1.0 -p dist/SESMG.deb
       - name: Upload Artifact
+        if: ${{ env.current_release != env.version }}
         uses: actions/upload-artifact@v3
         with:
           name: SESMG.deb
@@ -111,17 +129,21 @@ jobs:
       needs: [release, macos, linux]
       steps:
         - name: Checkout
+          if: ${{ env.current_release != env.version }}
           uses: actions/checkout@v2
 
         - name: Setup Python
+          if: ${{ env.current_release != env.version }}
           uses: actions/setup-python@v4
           with:
             python-version: '3.9.13'
             
         - name: Add msbuild to PATH
+          if: ${{ env.current_release != env.version }}
           uses: microsoft/setup-msbuild@v1.1
 
         - name: Create Executable
+          if: ${{ env.current_release != env.version }}
           run: |
             python3 -m venv .
             Scripts/activate.bat
@@ -132,6 +154,7 @@ jobs:
[…]
