+++ .github/workflows/semantic_versioning.yml
@@ -39,16 +39,16 @@ jobs:
         uses: actions/checkout@v2
       - name: get current version 
         run : |
-          version=$(python setup.py --version)
-          echo "version=v$version" >> $GITHUB_ENV
+          current_version=$(python setup.py --version)
+          echo "current_version=v$current_version" >> $GITHUB_ENV
           
       - name: Setup Python
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         run: |
           curl https://www.python.org/ftp/python/3.9.13/python-3.9.13-macosx10.9.pkg --output pythonInstaller.pkg
           sudo installer -pkg pythonInstaller.pkg -target /
       - name: Create Executable
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         run: |
           python3 -m venv .
           . bin/activate
@@ -57,7 +57,7 @@ jobs:
           pip3 install -r requirements.txt  
           pyinstaller .github/deploy/SESMG_macos.spec
       - name: Prepare for deployment
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         run: |
           brew install create-dmg
           mkdir -p dist/dmg
@@ -68,7 +68,7 @@ jobs:
           "dist/SESMG.dmg" \
           "dist/dmg/"
       - name: Upload Artifact
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         uses: actions/upload-artifact@v3
         with:
           name: SESMG.dmg
@@ -80,17 +80,17 @@ jobs:
     needs: [release, macos]
     steps:
       - name: Checkout
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         uses: actions/checkout@v2
 
       - name: Setup Python
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         uses: actions/setup-python@v4
         with:
           python-version: '3.9.13'
 
       - name: Create Executable
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         run: |
           pip3 install virtualenv
           python3 -m virtualenv .
@@ -103,7 +103,7 @@ jobs:
           pip3 install -r requirements.txt  
           pyinstaller .github/deploy/SESMG_linux.spec
       - name: Prepare for deployment
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         run: |
           sudo apt-get install ruby-dev build-essential && sudo gem i fpm -f
           mkdir -p package/opt
@@ -117,7 +117,7 @@ jobs:
           chmod +x package/opt/SESMG/SESMG
           fpm -C package -s dir -t deb -n "SESMG" -v 0.1.0 -p dist/SESMG.deb
       - name: Upload Artifact
-        if: ${{ env.current_release != env.version }}
+        if: ${{ env.current_release != env.current_version }}
         uses: actions/upload-artifact@v3
         with:
           name: SESMG.deb
@@ -129,21 +129,21 @@ jobs:
       needs: [release, macos, linux]
       steps:
         - name: Checkout
-          if: ${{ env.current_release != env.version }}
+          if: ${{ env.current_release != env.current_version }}
           uses: actions/checkout@v2
 
         - name: Setup Python
-          if: ${{ env.current_release != env.version }}
+          if: ${{ env.current_release != env.current_version }}
           uses: actions/setup-python@v4
           with:
             python-version: '3.9.13'
             
         - name: Add msbuild to PATH
[…]
