+++ .github/workflows/semantic_versioning.yml
@@ -74,7 +74,7 @@ jobs:
           
   linux:
     name: Build on linux
-    if: ${{ env.current_release != env.current_version }}
+    if: env.current_release != env.current_version
     runs-on: ubuntu-20.04
     needs: [release, macos]
     steps:
@@ -119,7 +119,7 @@ jobs:
   
   windows:
       name: Build on windows
-      if: ${{ env.current_release != env.current_version }}
+      if: env.current_release != env.current_version
       runs-on: windows-latest
       needs: [release, macos, linux]
       steps:
@@ -154,7 +154,7 @@ jobs:
             
   deploy:
     name: Push builds 
-    if: ${{ env.current_release != env.current_version }}
+    if: env.current_release != env.current_version
     runs-on: ubuntu-20.04
     needs: [release, linux, macos, windows]
     steps:
