+++ .github/workflows/semantic_versioning.yml
@@ -28,7 +28,7 @@ jobs:
         id: before_release_version
         run: |
           current_release=$(python3 setup.py --version)
-          echo "::set-env name=CURRENT_RELEASE::v$current_release"
+          echo "CURRENT_RELEASE::v$current_release" >> $GITHUB_ENV
 
       - name: Python Semantic Release
         uses: relekang/python-semantic-release@master
