+++ .github/workflows/semantic_versioning.yml
@@ -6,10 +6,6 @@ on:
       - master
       - test
       
-env:
-  CURRENT_VERSION: "v0.0.0"
-  CURRENT_RELEASE: "v0.0.0"
-      
 permissions: 
   contents: read
 
@@ -34,12 +30,22 @@ jobs:
         uses: relekang/python-semantic-release@master
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
+      
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
           
       - name: get current version 
         id: after_release_version 
         run : |
           current_version=$(python3 setup.py --version)
           echo "CURRENT_VERSION=v$current_version" >> $GITHUB_ENV
+          
+      - name: check rather a new verion was released       
+        run:  |
+         if ${{ env.CURRENT_RELEASE != env.CURRENT_VERSION }}; then
+            exit 0
+         fi
 
   macos:
     name: Build for macOS
@@ -51,9 +57,6 @@ jobs:
           
       - name: Setup Python
         run: |
-          echo $CURRENT_RELEASE
-          echo $CURRENT_VERSION
-          echo $CURRENT_RELEASE != $CURRENT_VERSION
           curl https://www.python.org/ftp/python/3.9.13/python-3.9.13-macosx10.9.pkg --output pythonInstaller.pkg
           sudo installer -pkg pythonInstaller.pkg -target /
       - name: Create Executable
