+++ .github/workflows/semantic_versioning.yml
@@ -43,7 +43,7 @@ jobs:
           
       - name: check rather a new verion was released       
         run:  |
-         if ${{ env.CURRENT_RELEASE != env.CURRENT_VERSION }}; then
+         if ${{ env.CURRENT_RELEASE == env.CURRENT_VERSION }}; then
             exit 0
          fi
 
