+++ .github/workflows/semantic_versioning.yml
@@ -44,11 +44,10 @@ jobs:
       - name: check rather a new verion was released       
         id: new_release
         run:  |
-         if ${{ env.CURRENT_RELEASE == env.CURRENT_VERSION } == 'true'} ; then
-            echo "bool = ${{ env.CURRENT_RELEASE != env.CURRENT_VERSION }}" >> $GITHUB_OUTPUT
-         fi
-         if ${{ env.CURRENT_RELEASE != env.CURRENT_VERSION } == 'true'}; then
-            echo "bool = ${{ env.CURRENT_RELEASE != env.CURRENT_VERSION }}" >> $GITHUB_OUTPUT
+         if ${{ env.CURRENT_RELEASE == env.CURRENT_VERSION }} ; then
+            echo "bool = 'false'" >> $GITHUB_OUTPUT
+         else
+            echo "bool = 'true'" >> $GITHUB_OUTPUT
          fi
 
   macos:
