+++ .github/workflows/semantic_versioning.yml
@@ -52,9 +52,9 @@ jobs:
 
   macos:
     name: Build for macOS
-    if: ${{ needs.release.new_release.outputs.bool }} == 'true'
     runs-on: macos-latest
     needs: release
+    if: ${{ needs.release.new_release.outputs.bool  == 'true' }}
     steps:
       - name: Checkout
         uses: actions/checkout@v2
