+++ .github/workflows/semantic_versioning.yml
@@ -15,6 +15,8 @@ jobs:
     concurrency: release
     permissions: 
       contents: write
+    outputs:
+      bool: ${{ steps.new_release.outputs.bool }}
     steps:
       - uses: actions/checkout@v2
         with:
@@ -60,7 +62,7 @@ jobs:
           
       - name: Setup Python
         run: |
-          echo ${{ needs.release.new_release.outputs.bool }} 
+          echo ${{ needs.release.outputs.bool }} 
           curl https://www.python.org/ftp/python/3.9.13/python-3.9.13-macosx10.9.pkg --output pythonInstaller.pkg
           sudo installer -pkg pythonInstaller.pkg -target /
       - name: Create Executable
