+++ .github/workflows/semantic_versioning.yml
@@ -56,7 +56,7 @@ jobs:
     name: Build for macOS
     runs-on: macos-latest
     needs: release
-    if: ${{ needs.release.outputs.bool }} 
+    if: ${{ needs.release.outputs.bool == 'true' }} 
     steps:
       - name: Checkout
         uses: actions/checkout@v2
