+++ .github/workflows/semantic_versioning.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - master
-      - test
       
 permissions: 
   contents: read
