### 2023-08-03 ce685e3
code(): add graphviz dependency check
