+++ program_files/GUI_st/1_Main_Application.py
@@ -556,6 +556,10 @@ def change_state_submitted_clear_cache() -> None:
 model_definition_input_file = main_input_sidebar()
 main_clear_cache_sidebar()
 
+# check rather the dependencies to be installed by the user are
+# installed
+GUI_functions.check_for_dependencies()
+
 # load the start page if modell run is not submitted
 if st.session_state["state_submitted_optimization"] == "not done":
     main_start_page()
