+++ program_files/start_script.py
@@ -1,7 +1,5 @@
-from program_files.GUI_st import GUI_st_global_functions
 import sys
 from streamlit.web import cli as stcli
-import streamlit
 
 from pathlib import Path
 
