### 2023-08-03 3a9bca7
Revert "code(main_application): test to raise additional error information"
This reverts commit 76c3910aa5d5164daa88aa9710b89394974db40b.
