### 2023-08-03 d0c0f3b
code(GUI):result table, error message for troubleshooting and additional graph when model crashes
