+++ program_files/GUI_st/pages/3_Result_Processing.py
@@ -13,7 +13,7 @@ from PIL import Image
 
 from program_files.GUI_st.GUI_st_global_functions import \
     import_GUI_input_values_json, st_settings_global, read_markdown_document, \
-    load_result_folder_list
+    load_result_folder_list, show_error_with_link
 
 
 def result_processing_sidebar() -> None:
@@ -206,6 +206,7 @@ def short_result_table(result_path_components: str) -> None:
     st.subheader("Result Table")
     # Import components.csv and create dataframe
     df_components = pd.read_csv(result_path_components)
+
     # CSS to inject contained in a string
     hide_dataframe_row_index = """
                 <style>
@@ -215,24 +216,43 @@ def short_result_table(result_path_components: str) -> None:
                 """
 
     # create table
+
+    # add dicctionarry for a possible empty row
+    # funktioniert hier nicht, da Java Skript mit null arbeitet statt mit None, könnte man transformieren, aber erstmal
+    # Jan fragen, ob es nicht auch leichter geht
+    leere_zeile = {}
+    for spalte in df_components.columns:
+        leere_zeile[spalte] = None
+
+    df_components = df_components.append(leere_zeile, ignore_index=True)
+
+    print(df_components)
+
+
+
     # set min hight which is the header height
     ag_min_height = 40
-    # set right per row
+    # set hight per row
     ag_row_height = 27.6
     # calculate logical height based on the df length
-    logical_df_height = ag_min_height + len(df_components) * ag_row_height
+    logical_df_height = ag_min_height + (len(df_components)+1) * ag_row_height
     # set maximum height for st.AgGrid Table
-    ag_max_height = 500
+    ag_max_height = 200
 
     # Inject CSS with Markdown
     st.markdown(hide_dataframe_row_index, unsafe_allow_html=True)
 
     # Creating st_aggrid table with setting the height to the min of the \
     # logical height or the max height
+
     AgGrid(df_components,
            fit_columns_on_grid_load=True,
-           height=min(logical_df_height, ag_max_height))
+           height=min(ag_max_height,logical_df_height))
 
+    # das Problem tritt jetzt noch auf wenn ag_max_height ausgewählt wird, man könnte natürlich auch einfach immer die
+    # ganze Tabelle anzeigen lassen, aber ich suche noch nach dem Grund warum es bei max height abgeschnitten wird
+    # Idee Benjamin war noch eine leere Zeile anzeigen zu lassen, habe ich versucht, funktioniert, aber auch nur wenn
+    # die logical height ausgewählt wird, siehe oben
 
 def short_result_interactive_dia(result_path_results: str) -> None:
     """
@@ -361,121 +381,133 @@ def short_result_graph(result_path_graph: str) -> None:
         st.image(es_graph)
 
 
-# starting page functions
-# initialize global page settings
-st_settings_global()
+try:
+    # starting page functions
+    # initialize global page settings
+    st_settings_global()
 
-# initialize session state  if no result paths are defined on main page
-if "state_result_path" not in st.session_state:
-    st.session_state["state_result_path"] = "not set"
+    # initialize session state  if no result paths are defined on main page
+    if "state_result_path" not in st.session_state:
+        st.session_state["state_result_path"] = "not set"
 
-# start sidebar functions
-result_processing_sidebar()
+    # start sidebar functions
+    result_processing_sidebar()
 
-# show introduction page if no result paths are not set
-if st.session_state["state_result_path"] == "not set":
-    read_markdown_document(
-        document_path="docs/GUI_texts/results.md",
-        folder_path=f'{"docs/images/manual/Results/*"}')
+    # show introduction page if no result paths are not set
+    if st.session_state["state_result_path"] == "not set":
[…]
