### 2023-08-03 362789c
code(Executable): add own hosted window by starting up the executable using pyside2
