### 2023-08-03 27771eb
Update 01.01.00_structure_of_energy_systems.rst
