### 2023-08-04 2ca089d
code(startscript): fix to be anaconda compatible
