### 2023-08-07 800ed79
code(GUI): switch GUI to new result folder structure (/user/documents/SESMG/results)
