### 2023-08-07 SESMG/SESMG#193 @willu47
This is okay to merge.  I noted one typo "easilyly" in the intro.
