### 2023-08-08 SESMG/SESMG#194 @nick-gorman
@chrklemm Apologies for my delayed response, these changes look good and address my concerns.
