### 2023-08-08 79ab24d
docs(paper.md): typo
