### 2023-08-08 SESMG/SESMG#193 @chrklemm
Thank you for your effort @willu47! I have corrected the typo and merged the changes into the master branch!
