+++ tests/test_GUI_st_GUI_st_global_functions.py
@@ -216,8 +216,7 @@ def test_positive_read_markdown_document():
 
     # creating the reduced README file as a str in a list
     reduced_readme = read_markdown_document(
-        document_path=os.path.join(
-            os.path.dirname(os.path.dirname(__file__)), "README.md"),
+        document_path="README.md",
         folder_path="empty")
 
     # iterate through the list and check if unique text part bevor and \
