+++ program_files/start_script.py
@@ -6,13 +6,19 @@ parent = os.path.abspath('..')
 sys.path.insert(1, parent)
 
 from program_files.GUI_st import GUI_st_global_functions
+import streamlit
+from streamlit.web import cli as stcli
 from pathlib import Path
 import atexit
 import subprocess as sp
+import matplotlib.pyplot as plt
+import matplotlib
 
+if getattr(sys, 'frozen', False) and sys.platform == 'darwin':
+    os.environ["QTWEBENGINE_RESOURCES_PATH"] = str(Path(sys._MEIPASS))
+    os.environ["QTWEBENGINE_LOCALES_PATH"] = str(Path(sys._MEIPASS)) + "/qtwebengine_locales"
 
-from PySide2 import QtCore, QtWebEngineWidgets, QtWidgets
-
+from PySide6 import QtCore, QtWebEngineWidgets, QtWidgets
 
 def kill_server(p):
     if os.name == 'nt':
@@ -25,13 +31,14 @@ def kill_server(p):
 
 
 if __name__ == '__main__':
+    matplotlib.use("QtAgg")
+    plt.rcParams['figure.hooks'].append('mplcvd:setup')
+
     if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
         bundle_dir = Path(sys._MEIPASS)
     else:
         bundle_dir = Path(__file__).parent.parent
     
-    #cmd = "streamlit run {} --server.headless=True".format(str(bundle_dir) + "/program_files/GUI_st/1_Main_Application.py")
-
     cmd = ["streamlit",
            "run",
            str(bundle_dir)
@@ -39,16 +46,15 @@ if __name__ == '__main__':
              "--server.headless=True",
              "--global.developmentMode=False"
     ]
-    
     p = sp.Popen(cmd, stdout=sp.DEVNULL)
     atexit.register(kill_server, p)
-                
+    
     hostname = 'localhost'
     port = 8501
-
+    
     app = QtWidgets.QApplication()
     view = QtWebEngineWidgets.QWebEngineView()
-
+    
     view.load(QtCore.QUrl(f'http://{hostname}:{port}'))
     view.show()
-    app.exec_()
+    app.exec()
