### 2023-08-08 c3099f2
code(GUI): add architecture to allow the user to set an individual result directory path
