### 2023-08-08 c2e9fb7
code(start_script): add comments and improve pep8 conformity
