### 2023-08-09 8025fdb
code(): update macos executable
