### 2023-08-09 c5daa0d
code(urban_upscaling): change udu and GUI to new result folder structure
