### 2023-08-09 791465d
code(GUI): fix help comment reference
