### 2023-08-10 SESMG/SESMG#188 @GregorBecker
Hello @willu47,

thanks for the recommendations of the python packages for creating executable applications. We decided to use the following solution:

1. We have set up our application as a Streamlit GUI in a Pyside6 environment, so that the process is terminated by closing the window and the process does not continue to run in the background unintentionally.

2. We created a Github action to generate system executables running on Pyinstaller, which on the one hand makes them available after each push in the pre-release (e. g. https://github.com/SESMG/SESMG/releases/tag/SESMG_beta) and also triggers them with each semantic release, so that each version also ships with its executables.

However, we are still facing some unexpected challenges (for example, concerning our application's structure in combination with an executable PySide6 interface) that we thought we could address more quickly. Therefore we would like to apologize for the delays and will inform you as soon as we solved the few remaining issues. You can see our development progress in the issue-188 branch (https://github.com/SESMG/SESMG/tree/issue-188). We hope to be done soon.

Kind regards
SESMG Dev Team
