### 2023-08-10 73d68ea
code(GUI): update GUI cache to new folder structure
