+++ program_files/GUI_st/GUI_st_global_functions.py
@@ -6,16 +6,24 @@
 import json
 import glob
 import os
-import streamlit as st
 import sys
 from pathlib import Path
+import streamlit as st
 
 from program_files.preprocessing.Spreadsheet_Energy_System_Model_Generator \
     import sesmg_main, sesmg_main_including_premodel
 
 
 def get_bundle_dir():
+    """
+       Get the path to the bundle directory.
 
+       This function determines and returns the path to the bundle directory
+       where the program is located. It considers both the frozen state and
+       the non-frozen state of the program.
+
+       :returns: - **bundle_dir** (str) - The path to the bundle directory.
+    """
     if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
         bundle_dir = Path(sys._MEIPASS)
     else:
@@ -52,18 +60,33 @@ def import_GUI_input_values_json(json_file_path: str) -> dict:
             exported dict from json file including a (sub)dict for \
             every GUI page
     """
-    # Import json file including several (sub)dicts for every GUI page
-    # Each (sub)dict includes input values as a cache from the last session
-    with open(json_file_path, "r", encoding="utf-8") as infile:
-        GUI_settings_cache_dict_reload = json.load(infile)
+
+    # load the blank GUI_st_cache.json if the user specific directory
+    # does not exist
+    if os.path.exists(json_file_path) is False:
+
+        # define path to blank json cache file
+        blank_json_file_path = \
+            os.path.dirname(__file__) + "/GUI_st_cache.json"
+        # Import json file including several (sub)dicts for every GUI page
+        # Each (sub)dict includes the default GUI settings
+        with open(blank_json_file_path, "r", encoding="utf-8") as infile:
+            GUI_settings_cache_dict_reload = json.load(infile)
+
+    # else load the user sprecific GUI settings
+    else:
+        # Import json file including several (sub)dicts for every GUI page
+        # Each (sub)dict includes input values as a cache from the last session
+        with open(json_file_path, "r", encoding="utf-8") as infile:
+            GUI_settings_cache_dict_reload = json.load(infile)
 
     return GUI_settings_cache_dict_reload
 
 
-def save_GUI_input_values(input_values_dict: dict,
-                          json_file_path: str) -> None:
+def save_GUI_cache_dict(input_values_dict: dict,
+                        json_file_path: str) -> None:
     """
-        Function to save a dict as json.
+        Function to save a the cache dict as json.
 
         :param input_values_dict: name of the dict of input values \
             for specific GUI page
@@ -72,6 +95,28 @@ def save_GUI_input_values(input_values_dict: dict,
             input values
         :type json_file_path: str
     """
+
+    # check if internal /.app directory exists
+    if os.path.exists(json_file_path) is False:
+
+        # create directory
+        os.makedirs(set_internal_directory_path())
+
+    # dump the json file to the given directory
+    save_json_file(input_values_dict=input_values_dict,
+                   json_file_path=json_file_path)
+
+
+def save_json_file(input_values_dict: dict, json_file_path: str) -> None:
+    """
+        Function to save a dict as json.
+
+        :param input_values_dict: name of the dict
+        :type input_values_dict: dict
+        :param json_file_path: file path to the underlying json
+        :type json_file_path: str
+    """
+    # dump cache.json file to internal .app directory
     with open(json_file_path, 'w', encoding="utf-8") as outfile:
         json.dump(input_values_dict, outfile, indent=4)
 
@@ -84,46 +129,20 @@ def clear_GUI_main_settings(json_file_path: str) -> None:
         :param json_file_path: internal path where json should be saved
         :type json_file_path: str
     """
[…]
