### 2023-08-10 6f24ec1
code(deployment): update windows executable
