### 2023-08-11 ecf6274
code(GUI): fix reference to help comment
