### 2023-08-14 8ea4122
code(deployment): update building process of macos executable
