+++ hooks/hook-oemof.db.py
@@ -1,0 +1,5 @@
+from PyInstaller.utils.hooks import collect_submodules
+from PyInstaller.utils.hooks import collect_data_files
+hiddenimports = collect_submodules('oemof.db')
+
+datas = collect_data_files('oemof.db', include_py_files=True)
