+++ hooks/hook-oemof.py
@@ -1,0 +1,5 @@
+from PyInstaller.utils.hooks import collect_submodules
+from PyInstaller.utils.hooks import collect_data_files
+hiddenimports = collect_submodules('oemof')
+
+datas = collect_data_files('oemof')
