+++ hooks/hook-dhnx.py
@@ -1,0 +1,6 @@
+from PyInstaller.utils.hooks import collect_submodules
+from PyInstaller.utils.hooks import collect_data_files
+hiddenimports = collect_submodules('dhnx')
+
+datas = collect_data_files('dhnx', include_py_files=True)
+datas += collect_data_files('dhnx', include_py_files=True)
