+++ requirements.txt
@@ -21,5 +21,4 @@ dash_canvas>=0.1.0
 h5py==3.6.0
 pandas<2.0.0
 pytest
-pyqt6
-pyside6
+pyside2
