### 2023-08-16 2e6ff87
docu(results): introduce new result folder structure
