+++ docs/02.02.04_results.rst
@@ -46,8 +46,14 @@ The difference is that you need to select a folder that you want to analyze.
 Results as Spreadsheets and Log-Files
 -------------------------------------
 
-The results of the modeling are stored in the "results" folder in two formats:
-- as summarizing log files, under
+The results of the modeling are stored in the SESMG result folder which is 
+places by dafault in die /users/documents/SESMG/result direcotory. The directory
+is created be the SESMG application. You can change the direcotry when you are 
+using the advanced installation by following the `note information in the installation
+description <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#advanced-installation>`_.
+
+The results are saved in two formats:
+- as summarizing log files
 - as detailed xlsx-files.
 
 The log-file gives an overview of which components are created and which of the
