+++ docs/02.02.04_results.rst
@@ -3,7 +3,12 @@ Results
 
 .. warning::
 
-	If a time series preparation algorithm was applied for the (main-)model, not every time-step of the model was modeled. This must be considered when analysing *energy amounts*. Simplifying, it can be assumed that the modeled amount of energy multiplied by the variable cost factor (see methods section) corresponds to the actual  amount of energy over the entire period. Changes in the cost structure, on the other hand, are automatically taken into account.
+	If a time series preparation algorithm was applied for the (main-)model, not 
+	every time-step of the model was modeled. This must be considered when 
+	analysing *energy amounts*. Simplifying, it can be assumed that the modeled 
+	amount of energy multiplied by the variable cost factor (see methods section) 
+	corresponds to the actual  amount of energy over the entire period. Changes 
+	in the cost structure, on the other hand, are automatically taken into account.
 
 .. _`interactive results`:
 
@@ -31,7 +36,12 @@ and all building-specific load profiles via the interactive results.
 
 1b. Multi-criteria optimization
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-In multi-criteria optimization, several scenarios are calculated. For more information take a look at the method: `Modeling Method <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/01.02.00_multi_criteria_optimization.html>`_. For each scenario, the results described for a single-criteria optimization can be displayed by selecting the reduction of the scenario (see (1) figure below). In addition, a Pareto diagram and energy amount diagrams are displayed (see (2) figure below).
+In multi-criteria optimization, several scenarios are calculated. For more 
+information take a look at the method: `Modeling Method <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/01.02.00_multi_criteria_optimization.html>`_. 
+For each scenario, the results described for a single-criteria optimization can 
+be displayed by selecting the reduction of the scenario (see (1) figure below). 
+In addition, a Pareto diagram and energy amount diagrams are displayed (see (2) 
+figure below).
 
 .. figure:: ../docs/images/manual/GUI/gui_result_pareto.png
    :width: 50 %
@@ -50,7 +60,7 @@ The results of the modeling are stored in the SESMG result folder which is
 places by dafault in die /users/documents/SESMG/result direcotory. The directory
 is created be the SESMG application. You can change the direcotry when you are 
 using the advanced installation by following the `note information in the installation
-description <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#advanced-installation>`_.
+description <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#extended-installation>`_.
 
 The results are saved in two formats:
 - as summarizing log files
