### 2023-08-16 c4b6d29
publication(paper.md): Update order of authors
