### 2023-08-16 f9bfa34
code(hooks): update pyinstaller hooks
