+++ hooks/hook-streamlit.py
@@ -5,3 +5,41 @@ hiddenimports += collect_submodules('st_aggrid')
 
 datas = collect_data_files('streamlit', include_py_files=True)
 datas += collect_data_files('st_aggrid', include_py_files=True)
+datas += collect_data_files('toml', include_py_files=True)
+datas += collect_data_files('blinker', include_py_files=True)
+datas += collect_data_files('click', include_py_files=True)
+datas += collect_data_files('packaging', include_py_files=True)
+datas += collect_data_files('importlib_metadata', include_py_files=True)
+datas += collect_data_files('zipp', include_py_files=True)
+datas += collect_data_files('tzlocal', include_py_files=True)
+datas += collect_data_files('pympler', include_py_files=True)
+datas += collect_data_files('rich', include_py_files=True)
+datas += collect_data_files('validators', include_py_files=True)
+datas += collect_data_files('tenacity', include_py_files=True)
+datas += collect_data_files('requests', include_py_files=True)
+datas += collect_data_files('pydeck', include_py_files=True)
+datas += collect_data_files('cachetools', include_py_files=True)
+datas += collect_data_files('toolz', include_py_files=True)
+datas += collect_data_files('jinja2', include_py_files=True)
+datas += collect_data_files('gitdb', include_py_files=True)
+datas += collect_data_files('pyparsing', include_py_files=True)
+datas += collect_data_files('pytz', include_py_files=True)
+datas += collect_data_files('urllib3', include_py_files=True)
+datas += collect_data_files('idna', include_py_files=True)
+datas += collect_data_files('pygments', include_py_files=True)
+datas += collect_data_files('smmap', include_py_files=True)
+datas += collect_data_files('mdurl', include_py_files=True)
+datas += collect_data_files('tzdata', include_py_files=True)
+datas += collect_data_files('google', include_py_files=True)
+datas += collect_data_files('pandas', include_py_files=True)
+datas += collect_data_files('dateutil', include_py_files=True)
+datas += collect_data_files('numpy', include_py_files=True)
+datas += collect_data_files('pyarrow', include_py_files=True)
+datas += collect_data_files('numexpr', include_py_files=True)
+datas += collect_data_files('PIL', include_py_files=True)
+datas += collect_data_files('pytz_deprecation_shim', include_py_files=True)
+datas += collect_data_files('tornado', include_py_files=True)
+datas += collect_data_files('matplotlib', include_py_files=True)
+datas += collect_data_files('charset_normalizer', include_py_files=True)
+datas += collect_data_files('certifi', include_py_files=True)
+datas += collect_data_files('colorama', include_py_files=True)
