### 2023-08-16 429d726
code(start_script.py): update start_script.py for windows pyinstaller executable
