### 2023-08-16 823c1e5
code(GUI): fix result start page, update text
