### 2023-08-16 de7ebcb
docu(results): typos
