+++ docs/02.02.04_results.rst
@@ -8,7 +8,8 @@ Results
 	analysing *energy amounts*. Simplifying, it can be assumed that the modeled 
 	amount of energy multiplied by the variable cost factor (see methods section) 
 	corresponds to the actual  amount of energy over the entire period. Changes 
-	in the cost structure, on the other hand, are automatically taken into account.
+	in the cost structure, on the other hand, are taken into account 
+	automatically.
 
 .. _`interactive results`:
 
@@ -57,8 +58,8 @@ Results as Spreadsheets and Log-Files
 -------------------------------------
 
 The results of the modeling are stored in the SESMG result folder which is 
-places by dafault in die /users/documents/SESMG/result direcotory. The directory
-is created be the SESMG application. You can change the direcotry when you are 
+places by dafault in the /users/documents/SESMG/result direcotory. The directory
+is created by the SESMG application. You can change the direcotry when you are 
 using the advanced installation by following the `note information in the installation
 description <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#extended-installation>`_.
 
