+++ hooks/hook-streamlit.py
@@ -43,3 +43,4 @@ datas += collect_data_files('matplotlib', include_py_files=True)
 datas += collect_data_files('charset_normalizer', include_py_files=True)
 datas += collect_data_files('certifi', include_py_files=True)
 datas += collect_data_files('colorama', include_py_files=True)
+datas += collect_data_files('altair', include_py_files=True)
