### 2023-08-17 8420dae
docu(installation): update to new application (DEV)
