+++ docs/02.01.00_installation.rst
@@ -1,258 +1,537 @@
 Installation
 ************
 
-If you have programming experience and if you want to set up your own virtual environment, you may use the `advanced installation <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#advanced>`_. This allows, for example, the use of Anaconda.
+There are two possible ways to install the SESMG application. We can choose 
+between the installation as an independent application, which is appropriate 
+for common users, and an installation for advanced users or developers. We 
+recommend the installation as an independent application for the beginning. 
+You can easily switch later on.
+
+We recommend the advanced installation which will create the SESMG with its
+folder structure on your machine allowing you to work on it or adjust settings. 
+You can follow the `extended installation instructions <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#extended-installation>`_ 
+below. If you have programming experience and want to set your own virtual 
+environment (e. g. allowing the usage of Anaconda enviroments) you can follow 
+`advanced installation instructions <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#advanced-installation>`_. 
 
 
+
+SESMG as Application
+====================
+
+Windows
+-------
+
+1. Installing Python (3.8 or 3.9)
+Windows users can download the preferred version here: `Python download page for windows <https://www.python.org/downloads/windows/>`_.
+
+.. note:: 
+
+	Make sure to use the installer referring to your windows version (e.g. 
+	"Windows x86-64 executable installer" for windows 64-bit version). You can 
+	check the build in bit version following `these steps <https://support.microsoft.com/en-us/windows/32-bit-and-64-bit-windows-frequently-asked-questions-c6ca9541-8dce-4d48-0415-94a3faa2e13d>`_.
+	
+.. note:: 
+
+	The SESMG is not compatible with the python version 3.9.7, due to a 
+	confliciting bug in this version to the subpackage streamlit. You need to 
+	install another pyhton version or update to another python version e. g. 
+	3.9.13, which is widely tested.
+
+.. figure:: images/manual/Installation/sesmg_installation_ms_1.png
+   :width: 40 %
+   :align: center
+   
+.. note:: 
+
+	Make sure to select "Add python to PATH" at the beginning of the Python installation.
+
+.. figure:: images/manual/Installation/sesmg_installation_ms_2.png
+   :width: 40 %
+   :align: center
+
+
+2. Installing Graphiz
+You kann download Graphiz here: `Graphiz download <https://graphviz.org/download/>`_
+
+.. figure:: images/manual/Installation/sesmg_installation_ms_4.png
+   :width: 40 %
+   :align: center
+
+.. note:: 
+
+	Make sure to select "Add Graphiz to system path for all users" at the 
+	beginning of the installation, otherwise you can use the default settings.
+
+.. figure:: images/manual/Installation/sesmg_installation_ms_5.png
+   :width: 40 %
+   :align: center
+
+
+3. Installing a solver
+You can download the free and open-source CBC-solver here: `cbc solver for windows <https://www.coin-or.org/download/binary/Cbc/>`_. 
+Alternatively you can uses the `direct download link <https://www.coin-or.org/download/binary/Cbc/Cbc-2.10-win64-msvc16-mdd.zip>`_
+
+.. figure:: images/manual/Installation/sesmg_installation_ms_6.png
+   :width: 40 %
+   :align: center
+
+The gurobi-solver may be used as an **alternative** to the CBC-solver. Be 
+aware, that the gurobi-solver is not open-source and a license is required. 
[…]
