### 2023-08-17 ebcbeb3
docu(installation): update MacOS, add know issue
