### 2023-08-17 ef62709
docu(installation): typo
