### 2023-08-17 031d962
docu(troubleshooting): add known MacOS pyinstaller issue
