+++ docs/03.00.00_trouble_shooting.rst
@@ -1,6 +1,15 @@
 Troubleshooting
 ***************
-During execution of the model generator, error messages and associated error messages may occur. 
+
+During the installation or execution of the model generator, error messages 
+and associated error messages may occur. 
+
+
+.. warning:: 
+    There is a know issue for MacOS users which are using the SESMG as an 
+    application which leads to problems opening the application itself. 
+    See Error I-009 below to check how to fix it or check the 
+    `MacOS installation instructions <https://spreadsheet-energy-system-model-generator.readthedocs.io/en/latest/02.01.00_installation.html#macos/>`_.
 
 
 Installation
@@ -100,6 +109,26 @@ Error I-008: No module named streamlit II
 **Possible Error Cause:** The SESMG, or especially the subpackage stramlit, is not compatible with the python version 3.9.7.   
 
 **Debugging:** Check if you have this version. If so, you need to install another version or update the version. 
+
+Error I-009: The application does not start after double-clicking (MacOS)
+-------------------------------------------------------------------------
+
+**Possible Error Cause I:** Due to a bug in a subpackage it is not possible yet to 
+open the SESMG as commum by double-clicking it on MacOS. 
+    
+**Debugging I:** You need to use the terminal to open the application. Navitage 
+to the directory by using the command inputs (e. g. cd) or right-click the 
+folder in which the application is placed (e. g. "desktop") and choose "new 
+terminal at the folder". Type in the command ``open SESMG.app``.
+
+**Possible Error Cause II:** Due to a bug in a subpackage the SESMG does not 
+kill all processes after closing the application on MacOS.
+    
+**Debugging II:** To kill all processes you need to use your taskmanager and stop 
+the "SESMG"-processes manually. Make sure to kill it after each session. 
+Otherwise you will not be able to reopen it by using the way described 
+above. We are working on it!
+
 
 Modeling
 ========
