### 2023-08-17 5e0c3fd
code(GUI): drop unused lines
