+++ program_files/GUI_st/pages/3_Result_Processing.py
@@ -369,10 +369,6 @@ st_settings_global()
 if "state_result_path" not in st.session_state:
     st.session_state["state_result_path"] = "not set"
 
-#st.write(read_markdown_document(
-#    document_path="docs/GUI_texts/results.md",
-#    folder_path=f'{"docs/images/manual/Results/*"}'))
-
 # start sidebar functions
 result_processing_sidebar()
 
