### 2023-08-17 ca5aad9
code(GUI): drop duplicate
