+++ hooks/hook-open_FRED.py
@@ -6,3 +6,4 @@ datas = collect_data_files('open_FRED', include_py_files=True)
 datas += collect_data_files('alembic', include_py_files=True)
 datas += collect_data_files('mako', include_py_files=True)
 datas += collect_data_files('geoalchemy2', include_py_files=True)
+datas += collect_data_files('markupsafe', include_py_files=True)
