### 2023-08-17 c42d52d
code(start_script.py): update start_script
