### 2023-08-17 5e15b7c
Update GUI_st_global_functions.py
