### 2023-08-17 3a1112f
code(GUI): deactivate cbc-/gurobi-solver check
