### 2023-08-17 2e192b1
code(pyinstaller_macos_executable): update to pyinstaller development branch version
