### 2023-08-17 c72a910
code(pyinstaller_hooks): add hook for sklearn
