### 2023-08-18 325b930
docu(installation): update MacOS installation
