### 2023-08-18 101ac85
code(GUI): deactivate dependencies check
