### 2023-08-18 29965a2
code(GUI): reimplement Graphviz dependencies check
