### 2023-08-21 346a7ff
code(sklearn_hook): remove bug
