+++ hooks/hook-sklearn.py
@@ -3,4 +3,4 @@ from PyInstaller.utils.hooks import collect_data_files
 hiddenimports = collect_submodules('sklearn')
 
 datas = collect_data_files('sklearn', include_py_files=True)
-datas += collect_data_files('sklearn.extra', include_py_files=True)
+datas += collect_data_files('sklearn_extra', include_py_files=True)
