### 2023-08-21 9d6df10
code(CompressionsHeatTransformer): enable connection to thermal network as high temperature source
