### 2023-08-21 7575e63
code(GUI): reactivate dependencies check
