### 2023-08-21 6a3e47b
code(hook-sklearn): add dependency of sklearn
