### 2023-08-22 2d1820b
code(GUI): update error message, fix dependencies check
