### 2023-08-24 3712628
code(pyinstaller): allow parallel tasks in pyside GUI
