+++ program_files/start_script.py
@@ -1,34 +1,25 @@
+import traceback
+os.chdir(os.path.dirname(__file__))
 import sys
 import os
-
+import multiprocessing
+multiprocessing.freeze_support()
+    
 # setting new system path to be able to refer to parent directories
 parent = os.path.abspath('..')
 sys.path.insert(1, parent)
 
 import subprocess as sp
-from multiprocessing import Process
-import matplotlib.pyplot as plt
-import streamlit
-import atexit
-import matplotlib
-import streamlit.web.bootstrap
+#import matplotlib.pyplot as plt
+#import streamlit
+#import atexit
+#import matplotlib
+#import sys
+#import streamlit.web.bootstrap
 from pathlib import Path
+#from threading import Thread
 
-
-# Set environment variables for frozen macOS applications using Qt WebEngine
-if getattr(sys, 'frozen', False) and sys.platform == 'darwin':
-    sys.path.append(str(os.path.dirname(os.path.dirname(
-        os.path.abspath(__file__)))))
-os.chdir(os.path.dirname(__file__))
 from PySide2 import QtCore, QtWebEngineWidgets, QtWidgets
-
-if getattr(sys, 'frozen', False) and sys.platform == 'darwin':
-    os.environ["QTWEBENGINEPROCESS_PATH"] = os.path.normpath(os.path.join(
-        sys._MEIPASS, 'PySide2', 'QT', 'lib',
-        'QtWebEngineCore.framework', 'Helpers', 'QtWebEngineProcess.app',
-        'Contents', 'MacOS', 'QtWebEngineProcess'
-        ))
-
 from program_files.GUI_st import GUI_st_global_functions
 
 
@@ -53,33 +44,68 @@ def kill_server(p):
         pass
 
 
-if __name__ == '__main__':
-    os.chdir(os.path.dirname(__file__))
-
-    
-    if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS') and sys.platform == "darwin":
-        cmd = "python3.9 -m streamlit run {} --server.headless=True --global.developmentMode=False --server.enableXsrfProtection=false".format(
-            "./program_files/GUI_st/1_Main_Application.py")
-    elif getattr(sys, 'frozen', False) and sys.platform == "win32":
-        cmd = "python -m streamlit run {} --server.headless=True --global.developmentMode=False --server.enableXsrfProtection=false".format(
-            "program_files\\GUI_st\\1_Main_Application.py")
-    else:
-        cmd = "streamlit run {} --server.headless=True".format(
-                "./GUI_st/1_Main_Application.py")
-        
-    p = sp.Popen(cmd.split(), stdout=sp.DEVNULL)
-    atexit.register(kill_server, p)
+def create_pyside_gui():
 
     hostname = 'localhost'
     port = 8501
-    
-    
+            
     # Initialize a Qt application and WebEngineView
     app = QtWidgets.QApplication()
     view = QtWebEngineWidgets.QWebEngineView()
+                
 
     # Load and show the Streamlit app in the WebEngineView
     view.load(QtCore.QUrl(f'http://localhost:8501'))
     view.show()
-
     app.exec_()
+    
+
+def start_streamlit():
+    from streamlit import config as _config
+    try:
+        print("TEST")
+        print(sys._MEIPASS)
+        os.chdir(os.path.dirname(__file__))
+        _config.set_option("server.headless", True)
+        _config.set_option("global.developmentMode", False)
+        _config.set_option("server.enableXsrfProtection", False)
+        _config.show_config()
+        import streamlit.web.bootstrap
[…]
