### 2023-08-24 c9bc768
code(GUI): update graphviz path for M2
