### 2023-08-24 b183b7c
code(GUI): update graphviz PATH for M2
