### 2023-08-24 f00962f
code(startscript): update streamlit config and config.toml
