+++ program_files/start_script.py
@@ -1,5 +1,4 @@
 import traceback
-os.chdir(os.path.dirname(__file__))
 import sys
 import os
 import multiprocessing
@@ -18,6 +17,8 @@ import subprocess as sp
 #import streamlit.web.bootstrap
 from pathlib import Path
 #from threading import Thread
+
+os.chdir(os.path.dirname(__file__))
 
 from PySide2 import QtCore, QtWebEngineWidgets, QtWidgets
 from program_files.GUI_st import GUI_st_global_functions
@@ -48,64 +49,51 @@ def create_pyside_gui():
 
     hostname = 'localhost'
     port = 8501
-            
+
     # Initialize a Qt application and WebEngineView
     app = QtWidgets.QApplication()
     view = QtWebEngineWidgets.QWebEngineView()
-                
 
     # Load and show the Streamlit app in the WebEngineView
     view.load(QtCore.QUrl(f'http://localhost:8501'))
     view.show()
     app.exec_()
-    
+
 
 def start_streamlit():
     from streamlit import config as _config
-    try:
-        print("TEST")
-        print(sys._MEIPASS)
-        os.chdir(os.path.dirname(__file__))
-        _config.set_option("server.headless", True)
-        _config.set_option("global.developmentMode", False)
-        _config.set_option("server.enableXsrfProtection", False)
-        _config.show_config()
-        import streamlit.web.bootstrap
-        streamlit.web.bootstrap.run(str(sys._MEIPASS) + "/program_files/GUI_st/1_Main_Application.py", '', [], {})
-    except:
-        open("/Users/Shared/error1.log", "w").write(traceback.format_exc())
+    os.chdir(os.path.dirname(__file__))
+    _config.set_option("server.headless", True)
+    _config.set_option("global.developmentMode", False)
+    _config.set_option("server.enableXsrfProtection", False)
+    _config.show_config()
+    import streamlit.web.bootstrap
+    streamlit.web.bootstrap.run(str(sys._MEIPASS) + "/program_files/GUI_st/1_Main_Application.py", '', [], {})
 
 if __name__ == '__main__':
-    try:
-        #os.chdir(os.path.dirname(__file__))
-        
-        if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
-            multiprocessing.set_start_method("fork")
-            process2 = multiprocessing.Process(target=start_streamlit, args=[])
-            process2.start()
-            create_pyside_gui()
-             
-            #sys.argv = ["streamlit", "run", "--server.headless=True", "--global.developmentMode=False", "--server.enableXsrfProtection=false",  "./program_files/GUI_st/1_Main_Application.py"]
-            #sys.exit(stcli.main())
-           # if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS') and sys.platform == "darwin":
-           #     cmd = "python3.9 -m streamlit run {} --server.headless=True --global.developmentMode=False --server.enableXsrfProtection=false".format(
-           #         "./program_files/GUI_st/1_Main_Application.py")
-           # elif getattr(sys, 'frozen', False) and sys.platform == "win32":
-           #     cmd = "python -m streamlit run {} --server.headless=True --global.developmentMode=False --server.enableXsrfProtection=false".format(
-           #         "program_files\\GUI_st\\1_Main_Application.py")
-           # else:
-           #     cmd = "streamlit run {} --server.headless=True".format(
-           #             "./GUI_st/1_Main_Application.py")
-            
[…]
