### 2023-08-25 0208100
code(start_script): enable multiprocessing for win
