### 2023-08-25 2c5f684
code(start_script): typo if command
