### 2023-08-25 fff0bda
code(start_script): implement workaround for multiprocessing bug
