### 2023-08-25 2b141b5
code(GUI): add example win cbc PATH for testing
