### 2023-09-04 40e9682
ci(build.yml): set node.js version to the latest available version
