### 2023-09-04 5d38a9a
code(GUI_st_global_functions): add linux paths
