### 2023-09-04 bcc79a6
ci(build.yml): update checkout to v3
