### 2023-09-04 4eeccf8
code(start_script): partly fix macOS window layer problem, improve pep8 conformity
