### 2023-09-05 99a980a
code(startscript): remove mac options
