### 2023-09-05 8e24719
code(GUI_global_functions): fix gurobi for Mac M1
