### 2023-09-06 be2e77d
code(GUI_global_functions): add solver PATH definition for win
