### 2023-09-06 8b71fbf
code(GUI_global_functions): clean code and fix bug in win path definition
