+++ program_files/GUI_st/GUI_st_global_functions.py
@@ -413,110 +413,62 @@ def check_for_dependencies(solver: str):
     cbc_bool = False
     gurobi_bool = False
 
-
-    # Check if the selected solver is "cbc" and the current platform is not Windows.
-    if solver == "cbc" and not sys.platform == "win32":
-        # Use the 'which' command to find the path to the CBC solver executable.
-        cbc_path_str = os.popen('which cbc').read()[:-5]
-    
+    # Check if the selected solver is "cbc"
+    if solver == "cbc":
+        # Execution path for Mac and Linux platforms.
+        if not sys.platform == "win32":
+            # 'which' command to find the path to the CBC solver executable.
+            cbc_path_str = os.popen('which cbc').read()[:-5]
+        # Execution path for Mac and Windows platforms.
+        else:
+            # Use the 'where' command to find the path(s) to the CBC solver
+            # executable(s).
+            cbc_path_list = os.popen('where cbc').read()[:-5]
+            # Iterate through the list of paths to find the first valid one.
+            for i in cbc_path_list:
+                pos = i.find('.')
+                if pos > 0:
+                    # Extract and store the path without the file extension.
+                    i = i[0:pos]
+                    cbc_path_str = i
+                    break
         # Append the CBC solver path to the system's PATH environment variable.
         os.environ["PATH"] += os.pathsep + cbc_path_str
-    
+        print("CBC:")
+        print(cbc_bool)
+        print(cbc_path_str)
         # Set a boolean flag indicating that CBC solver is available.
         cbc_bool = True
-    
-    # Check if the selected solver is "gurobi" and the current platform is not Windows.
-    elif solver == "gurobi" and not sys.platform == "win32":
-        # Use the 'which' command to find the path to the Gurobi solver executable.
-        gurobi_path_str = os.popen('which gurobi.sh').read()[:-11]
-    
-        # Append the Gurobi solver path to the system's PATH environment variable.
+
+    # check is solver is gurobi
+    elif solver == "gurobi":
+
+        # Execution path for Mac and Linux platforms.
+        if not sys.platform == "win32":
+            # Use the 'which' command to find the path to the Gurobi solver
+            # executable.
+            gurobi_path_str = os.popen('which gurobi.sh').read()[:-11]
+        # Execution path for Windows platform.
+        else:
+            # 'where' command to find the path(s) to the Gurobi solver
+            # executable.
+            gurobi_path_list = os.popen('where gurobi').read()[:-5]
+            # Iterate through the list of paths to find the first valid one.
+            for i in gurobi_path_list:
+                pos = i.find('.')
+                if pos > 0:
+                    # Extract and store the path without the file extension.
+                    i = i[0:pos]
+                    gurobi_path_str = i
+                    break
+        # Append the Gurobi solver path to the system's PATH environment.
         os.environ["PATH"] += os.pathsep + gurobi_path_str
-    
+        print("gurobi:")
+        print(gurobi_bool)
+        print(gurobi_path_str)
         # Set a boolean flag indicating that Gurobi solver is available.
         gurobi_bool = True
-    
-    # If the solver is "cbc" and the platform is Windows, attempt to find the CBC solver executable.
-    elif solver == "cbc":
-        # Use the 'where' command to find the path(s) to the CBC solver executable(s).
-        cbc_path_list = os.popen('where cbc').read()[:-5]
-    
-        # Iterate through the list of paths to find the first valid one.
-        for i in cbc_path_list:
-            pos = i.find('.')
-            if pos > 0:
-                # Extract and store the path without the file extension.
-                i = i[0:pos]
-                cbc_path_str = i
-                break
-        
-        # Append the Gurobi solver path to the system's PATH environment variable.
[…]
