### 2023-09-06 b3c5d1c
code(GUI_global): fix bug in win path definition
