### 2023-09-06 5a6e849
docs(installation): typo
