### 2023-09-06 199b131
docs(installation): typo
