### 2023-09-06 23bc319
code(deployment): update pyside version from 2 to 6 to enable usage of plotlyjs
