### 2023-09-06 105d684
docs(installation): typo note boxes
