### 2023-09-06 67eb348
code(GUI_global_functions): add docstring
