+++ docs/02.01.00_installation.rst
@@ -87,13 +87,13 @@ the gurobi-solver is described in detail `here <https://www.gurobi.com/documenta
 The latest released SESMG version can be accessed here: `SESMG releases on Github <https://github.com/SESMG/SESMG/releases/latest/>`_. 
 Download the SESMG.exe and place it in a directory of your choice. Open the SESMG application by running SESMG.exe. The starting process usually takes a few minutes.
 
-.. note::
+.. tip::
 
 	With the first use you may have to confirm in windows-defender or other protection software that the SESMG is trustworthy. 
 	The exact warning message may differ depending on the system language and anti-virus software.
 
 
-.. note::
+.. tip::
 
 	While the SESMG is starting, a message like "the website is not available" may appear. Wait a moment and press "reload". If necessary, repeat this step. 
 	The exact warning message may differ depending on the system language and applied browser.
