### 2023-09-06 db68383
docs(installation): updated cbc installation steps
