### 2023-09-06 4ee8818
docs(installation): typo
