+++ docs/02.01.00_installation.rst
@@ -76,10 +76,11 @@ You kann download Graphiz here: `Graphiz download <https://graphviz.org/download
 3. Installing a solver
 
 3a. The free and open-source CBC-solver is recommended for new users and users which are not facing any run-time issues. To use the solver, the following steps must be carried out:
+
 	- if not already existing, create a folder called "SESMG" in the following directory: "/users/<your username>/documents/"
 	- download the most recent CBC-version for your operating system: `cbc solver for windows <https://www.coin-or.org/download/binary/Cbc/>`_. Alternatively you can uses the `direct download link 		  <https://www.coin-or.org/download/binary/Cbc/Cbc-2.10-win64-msvc16-mdd.zip>`_
 	- unzip the downloaded folder
-	- copy the "cbc.exe" (stored in the "bin" directory of the unziped folder) file into the following directory: ""/users/<your username>/documents/SESMG"
+	- copy the "cbc.exe" (stored in the "bin" directory of the unziped folder) file into the following directory: "/users/<your username>/documents/SESMG"
 
 3b. The gurobi-solver may be used as an **alternative** to the CBC-solver. Be 
 aware, that the gurobi-solver is not open-source and a license is required. 
