### 2023-09-06 cbf3208
code(GUI_global_functions): add path check for graphviz
