### 2023-09-07 330dffe
code(GUI_global): fix win dependencies check
