### 2023-09-07 bd7369c
docu(installation): drop python installation, add linux cmd commands
