### 2023-09-07 2265bda
code(GUI_global_function): remove development prints
