### 2023-09-07 3c714ac
code(GUI): try to fix image path for win
