### 2023-09-07 2bf5602
code(GUI): add test prints for win path
