### 2023-09-07 c43dff6
code(GUI_global_functions): complete docstring
