### 2023-09-11 669dd38
code(spec_files): fix windows paths to images
