### 2023-09-11 5f37c24
code(GUI): try bugfix image as url
