### 2023-09-12 023b243
code(GUI): fix image path
