### 2023-09-12 e415f24
docs(paper.md): typo
