### 2023-09-12 e8bd600
code(GUI): update demo tool image and text load
