### 2023-09-12 e919247
code(GUI_global_functions): update image loading to PIL
