### 2023-09-12 5fa9358
code(GUI_st_global_functions.py): add windows cbc path to the system PATH variable
