### 2023-09-12 addb37b
docs(installation): clarifications regarding windows installation
