### 2023-09-12 fdc1dbe
docu(02.01.00_installation.rst): update description of the installation procedure: Windows Coin Or Optimization Suite, Linux warning
