+++ docs/02.01.00_installation.rst
@@ -23,7 +23,7 @@ Windows
 -------
 
 1. Installing Graphiz
-You kann download Graphiz here: `Graphiz download <https://graphviz.org/download/>`_. Restart your computer after succesfully installing graphviz.
+You kann download Graphiz here: `Graphiz download <https://graphviz.org/download/>`_. 
 
 .. figure:: images/manual/Installation/sesmg_installation_ms_4.png
    :width: 40 %
@@ -34,10 +34,6 @@ You kann download Graphiz here: `Graphiz download <https://graphviz.org/download
 	Make sure to select "Add Graphiz to system path for all users" at the 
 	beginning of the installation, otherwise you can use the default settings.
 
-.. note:: 
-
-	Make sure to restart your computer before proceeding with the next steps.
-
 
 .. figure:: images/manual/Installation/sesmg_installation_ms_5.png
    :width: 40 %
@@ -46,18 +42,18 @@ You kann download Graphiz here: `Graphiz download <https://graphviz.org/download
 
 2. Installing a solver
 
-2a. The free and open-source CBC-solver is recommended for new users and users which are not facing any run-time issues. To use the solver, the following steps must be carried out:
-
-	- if not already existing, create a folder called "SESMG" in the following directory: "/documents/"
-	- download the most recent CBC-version for your operating system: `cbc solver for windows <https://www.coin-or.org/download/binary/Cbc/>`_. Alternatively you can uses the `direct download link 		  <https://www.coin-or.org/download/binary/Cbc/Cbc-2.10-win64-msvc16-mdd.zip>`_
-	- unzip the downloaded folder
-	- copy the "cbc.exe" (stored in the "bin" directory of the unziped folder) file into the following directory: "/documents/SESMG"
+2a. The free and open-source CBC-solver is recommended for new users and users which are not facing any run-time issues. To use the solver, download the most recent Coin-or-OptimizationSuite-version for your operating system: `Coin-or-OptimizationSuite for windows <https://www.coin-or.org/download/binary/OptimizationSuite/>`_. Alternatively you can uses the `direct download link 		  <https://www.coin-or.org/download/binary/OptimizationSuite/COIN-OR-1.8.0-win32-msvc12.exe>`_. Restart your computer after succesfully installing the solver.
 
 2b. The gurobi-solver may be used as an **alternative** to the CBC-solver. Be 
 aware, that the gurobi-solver is not open-source and a license is required. 
 However, the gurobi-solver may be faster for certain types of optimization problems, 
 especially for those containing binary decisions. The installation procedure of 
-the gurobi-solver is described in detail `here <https://www.gurobi.com/documentation/quickstart.html>`_.  
+the gurobi-solver is described in detail `here <https://www.gurobi.com/documentation/quickstart.html>`_. Restart your computer after succesfully installing the solver. 
+
+.. note:: 
+
+	Make sure to restart your computer before proceeding with the next steps.
+
 
 
 3. Installing the SESMG application
@@ -196,18 +192,19 @@ be sure the process startet by looking for SESMG.
 
 Use the following two commands in your terminal after installing the SESMG.
 
-``
-sudo apt install -y \
-libfontconfig1-dev libfreetype6-dev \
-libx11-dev libx11-xcb-dev libxext-dev libxfixes-dev \
-libxi-dev libxrender-dev \
-libxkbcommon-dev libxkbcommon-x11-dev libatspi2.0-dev \
-libopengl0 '^libxcb.*-dev' libgl1 \
-libegl1 \
-libdbus-glib-1-2``
[…]
