### 2023-09-12 b741683
docs(paper.md): typo
