### 2023-09-14 SESMG/SESMG#188 @willu47
Hi @chrklemm - congratulations. I have been able to test the installation for Mac OSX and can confirm that the installation process works as documented. It's a smooth experience.  It's a shame that you've got the license incompatibility problem, but there's not much you can do about that without changing your software stack.

I appreciate that this has been a longer-than-expected development process, but I think that it better meets your aims of providing a smooth user-facing experience for those who are not yet comfortable dealing with management of Python environments and batch or shell scripts.

Will
