### 2023-09-14 96ef348
major(semantic_versioning.yml): update the creation process of the SESMG executables within the semantic versioning
