### 2023-09-14 eaed3c2
feat(semantic_versioning.yml): update semantic versioning workflow
BREAKING CHANGE:
