### 2023-09-14 #5519 @chrklemm
Dear @timtroendle,

The proposed changes to the last open issue have been approved by the responsible reviewer @willu47. Thus, all suggested changes have been included in the SESMG and approved by the respective reviewers.

If further changes are needed, please let us know. However, for the case that no further changes are needed, we have already initiated the next steps:
- We have released a new version of the SESMG: v1.0.0
- The current version has been published on Zenodo (metadata identical to the Joss-paper) and can be found at the following DOI: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8344791.svg)](https://doi.org/10.5281/zenodo.8344791).

Once again, we would like to thank all the editors and reviewers involved for their help and patience during this review process.
Best regards,
SESMG-Dev-Team
